refactor(control): split controller.go by responsibility domain to eliminate god object
Split the 3988-line controller.go into 8 files by single responsibility:
- goal.go: Goal lifecycle management (~511 lines)
- approval.go: Approval/Permission system (~532 lines)
- checkpoint_ops.go: Checkpoint/Rewind/Fork (~294 lines)
- mcp.go: MCP Plugin management (~240 lines)
- snapshot.go: Snapshot/Autosave/Session destroy (~231 lines)
- memory_ops.go: Memory management (~116 lines)
- plan_todo.go: Plan/Todo state (~271 lines)
- controller.go: Turn execution core + Session lifecycle (~1864 lines)

No API changes — only method definitions moved to separate files within
the same package. This reduces bug density per file (controller.go had
21 bug fixes, the highest in the project) and makes each responsibility
domain independently reviewable and maintainable.
