+++ internal/control/checkpoint_ops.go
@@ -1,0 +1,294 @@
+package control
+
+import (
+	"context"
+	"fmt"
+	"log/slog"
+	"strconv"
+	"strings"
+
+	"reasonix/internal/agent"
+	"reasonix/internal/checkpoint"
+	"reasonix/internal/event"
+	"reasonix/internal/provider"
+)
+
+func ckptDir(sessionPath string) string {
+	if sessionPath == "" {
+		return ""
+	}
+	return strings.TrimSuffix(sessionPath, ".jsonl") + ".ckpt"
+}
+
+// rebindCheckpoints points the store at the (possibly new) session, loading any
+// checkpoints already on disk, and resets the turn boundaries. Called on
+// construction and whenever the session path changes (NewSession/Resume/SetSessionPath).
+func (c *Controller) rebindCheckpoints(sessionPath string) {
+	c.mu.Lock()
+	defer c.mu.Unlock()
+	c.goalStatePath = goalStatePath(sessionPath)
+	c.cp = checkpoint.New(ckptDir(sessionPath), c.cpRoot)
+	c.cpTurn = c.cp.NextTurn() // continue numbering past any checkpoints on disk
+	c.cpBound = c.cp.Bounds()  // rebuilt from persisted checkpoints so a resumed
+	if c.cpBound == nil {      // session can still rewind conversation / fork
+		c.cpBound = map[int]int{}
+	}
+}
+
+// beginCheckpoint opens a checkpoint for the turn about to run, recording the
+// current message count as the conversation-rewind boundary. Called at the top of
+// runTurn, before the user message is appended.
+func (c *Controller) beginCheckpoint(input string) {
+	if c.cp == nil || c.executor == nil {
+		return
+	}
+	c.mu.Lock()
+	turn := c.cpTurn
+	c.cpTurn++
+	msgIndex := len(c.executor.Session().Messages)
+	c.cpBound[turn] = msgIndex
+	c.mu.Unlock()
+	c.cp.Begin(turn, input, msgIndex)
+}
+
+func (c *Controller) Checkpoints() []checkpoint.Meta {
+	if c.cp == nil {
+		return nil
+	}
+	return c.cp.List()
+}
+
+// rewindFail emits the error as a Warn notice (so a frontend that swallows the
+// returned error — e.g. the desktop bridge's .catch — still shows the user why
+// the rewind did nothing) and returns it.
+func (c *Controller) rewindFail(err error) error {
+	c.sink.Emit(event.Event{Kind: event.Notice, Level: event.LevelWarn, Text: err.Error()})
+	return err
+}
+
+// Rewind restores the session to the start of `turn`: Code reverts every file that
+// turn (or a later one) changed to its pre-turn content; Conversation truncates the
+// message log back to that turn; Both does both. Refused while a turn is running.
+// Conversation rewind relies on the live boundary recorded at turn start, so it is
+// unavailable for turns inherited from a resumed session (code rewind still works).
+// Frontends re-render their transcript from History after the call.
+func (c *Controller) Rewind(turn int, scope RewindScope) error {
+	if c.cp == nil || c.executor == nil {
+		return c.rewindFail(fmt.Errorf("checkpoints unavailable"))
+	}
+	c.mu.Lock()
+	running := c.running
+	boundary, hasBound := c.cpBound[turn]
+	c.mu.Unlock()
+	if running {
+		return c.rewindFail(fmt.Errorf("cannot rewind while a turn is running"))
+	}
+
+	if scope == RewindCode || scope == RewindBoth {
+		written, deleted, err := c.cp.RestoreCode(turn)
+		if err != nil {
+			return c.rewindFail(fmt.Errorf("rewind code: %w", err))
+		}
+		if len(written) > 0 || len(deleted) > 0 {
+			c.sink.Emit(event.Event{Kind: event.Notice, Level: event.LevelInfo,
+				Text: fmt.Sprintf("rewound code to turn %d — %d file(s) restored, %d removed", turn, len(written), len(deleted))})
+		}
+	}
+	if scope == RewindConversation || scope == RewindBoth {
+		if !hasBound {
+			return c.rewindFail(fmt.Errorf("conversation rewind unavailable for turn %d (resumed session)", turn))
+		}
+		s := c.executor.Session()
+		// boundary is the message-log index at turn start; compaction shrinks the
+		// log without rewriting boundaries, so a stale boundary past the end means
+		// the turn was compacted away — fail loudly instead of skipping silently.
+		if boundary > len(s.Messages) {
+			return c.rewindFail(fmt.Errorf("conversation rewind unavailable for turn %d: the conversation was compacted past this point", turn))
+		}
+		s.Messages = s.Messages[:boundary]
+		c.mu.Lock()
+		c.cpTurn = turn // renumber future turns from here; later turns are gone
+		for k := range c.cpBound {
+			if k >= turn {
+				delete(c.cpBound, k)
+			}
+		}
+		c.mu.Unlock()
+		if err := c.Snapshot(); err != nil {
+			slog.Warn("controller: snapshot after rewind", "err", err)
+		}
+		c.sink.Emit(event.Event{Kind: event.Notice, Level: event.LevelInfo,
+			Text: fmt.Sprintf("rewound conversation to turn %d", turn)})
+	}
+	return nil
+}
+
+// Fork branches the conversation at the start of turn into a NEW session file,
+// preserving the current one as the branch point, and switches to the branch. Code
+// is untouched (it's a conversation operation). Like a conversation rewind it needs
+// the live boundary, so it is unavailable for resumed-session turns and refused
+// while a turn runs. Returns the new session path.
+func (c *Controller) Fork(turn int) (string, error) {
+	return c.ForkNamed(turn, "")
+}
+
+func (c *Controller) ForkNamed(turn int, name string) (string, error) {
+	return c.forkNamed(turn, name, true)
+}
+
+// ForkSession copies the conversation at the start of turn into a new session
+// file without switching this controller to it. Desktop uses this to open the
+// branch in a new tab while the source tab keeps its current transcript.
+func (c *Controller) ForkSession(turn int, name string) (string, error) {
+	return c.forkNamed(turn, name, false)
+}
+
+func (c *Controller) forkNamed(turn int, name string, switchToFork bool) (string, error) {
+	if c.executor == nil {
+		return "", c.rewindFail(fmt.Errorf("checkpoints unavailable"))
+	}
+	if c.sessionDir == "" {
+		return "", c.rewindFail(fmt.Errorf("fork needs session persistence, which is disabled"))
+	}
+	c.mu.Lock()
+	running := c.running
+	boundary, hasBound := c.cpBound[turn]
+	c.mu.Unlock()
+	if running {
+		return "", c.rewindFail(fmt.Errorf("cannot fork while a turn is running"))
+	}
+	if !hasBound {
+		return "", c.rewindFail(fmt.Errorf("fork unavailable for turn %d (resumed session)", turn))
+	}
+
+	// Persist the current conversation first so the branch point survives, then
+	// seed a fresh session with the messages up to the fork and switch to it.
+	if err := c.Snapshot(); err != nil {
+		slog.Warn("controller: pre-fork snapshot", "err", err)
+	}
+	parentPath := c.SessionPath()
+	parentID := agent.BranchID(parentPath)
+	src := c.executor.Session().Snapshot()
+	if boundary > len(src) {
+		boundary = len(src)
+	}
+	forked := append([]provider.Message(nil), src[:boundary]...)
+	sess := agent.NewSession("")
+	sess.Messages = forked
+
+	newPath := agent.NewSessionPath(c.sessionDir, c.label)
+	if err := sess.Save(newPath); err != nil {
+		return "", c.rewindFail(err)
+	}
+	if err := agent.SaveBranchMeta(newPath, agent.BranchMeta{
+		Name:             strings.TrimSpace(name),
+		ParentID:         parentID,
+		ForkTurn:         turn,
+		ForkMessageIndex: boundary,
+	}); err != nil {
+		return "", c.rewindFail(err)
+	}
+	if switchToFork {
+		c.executor.SetSession(sess)
+		c.resetPlannerSession()
+		c.mu.Lock()
+		c.sessionPath = newPath
+		c.mu.Unlock()
+		c.setActiveJobSession(newPath)
+		c.rebindCheckpoints(newPath)
+	}
+	c.sink.Emit(event.Event{Kind: event.Notice, Level: event.LevelInfo,
+		Text: fmt.Sprintf("forked conversation at turn %d into a new session", turn)})
+	return newPath, nil
+}
+
+func (c *Controller) CheckpointHasBoundary(turn int) bool {
+	c.mu.Lock()
+	defer c.mu.Unlock()
+	boundary, ok := c.cpBound[turn]
+	if !ok {
+		return false
+	}
+	// After compaction the key may still exist but the boundary value is
+	// stale (it points past the truncated message log).  Treat those
+	// turns the same as "no boundary" so the UI can disable the button.
+	return boundary <= len(c.executor.Session().Messages)
+}
+
+func (c *Controller) SummarizeFrom(ctx context.Context, turn int) error {
+	return c.summarizeAt(ctx, turn, true)
+}
+
+func (c *Controller) SummarizeUpTo(ctx context.Context, turn int) error {
+	return c.summarizeAt(ctx, turn, false)
+}
+
+func (c *Controller) summarizeAt(ctx context.Context, turn int, from bool) error {
+	if c.executor == nil {
+		return c.rewindFail(fmt.Errorf("checkpoints unavailable"))
+	}
+	c.mu.Lock()
+	running := c.running
+	boundary, hasBound := c.cpBound[turn]
+	c.mu.Unlock()
+	if running {
+		return c.rewindFail(fmt.Errorf("cannot summarize while a turn is running"))
+	}
+	if !hasBound {
+		return c.rewindFail(fmt.Errorf("summarize unavailable for turn %d (resumed session)", turn))
+	}
+	var err error
+	if from {
+		err = c.executor.SummarizeFrom(ctx, boundary)
+	} else {
+		err = c.executor.SummarizeUpTo(ctx, boundary)
+	}
+	if err != nil {
+		return c.rewindFail(err)
+	}
+	// The log was restructured; existing boundaries no longer map. Drop them (keep
+	// cpTurn monotonic so new turns don't collide with the store) — conversation
+	// rewind degrades to "unavailable" until fresh turns rebuild boundaries.
+	c.mu.Lock()
+	c.cpBound = map[int]int{}
+	c.mu.Unlock()
+	if err := c.Snapshot(); err != nil {
+		slog.Warn("controller: post-summarize snapshot", "err", err)
+	}
+	return nil
+}
+
+// parseRewind parses the arguments after "/rewind". The user may provide:
+//
+//	/rewind              → latest checkpoint, both
+//	/rewind <turn>       → that turn, both
+//	/rewind <turn> <scope> → that turn, code|conversation|both
+//
+// If no turn is given, the latest checkpoint is used. If no scope is given, Both is assumed.
+func parseRewind(args string, cps []checkpoint.Meta) (int, RewindScope, error) {
+	fields := strings.Fields(args)
+	if len(fields) == 0 {
+		if len(cps) == 0 {
+			return 0, RewindBoth, fmt.Errorf("no checkpoints available")
+		}
+		return cps[len(cps)-1].Turn, RewindBoth, nil
+	}
+	turn, err := strconv.Atoi(fields[0])
+	if err != nil {
+		return 0, RewindBoth, fmt.Errorf("invalid turn: %w", err)
+	}
+	scope := RewindBoth
+	if len(fields) >= 2 {
+		switch strings.ToLower(fields[1]) {
+		case "code":
+			scope = RewindCode
+		case "conversation":
+			scope = RewindConversation
+		case "both":
+			scope = RewindBoth
+		default:
+			return 0, RewindBoth, fmt.Errorf("unknown scope %q", fields[1])
+		}
+	}
+	return turn, scope, nil
+}
