+++ internal/control/controller.go
@@ -22,12 +22,10 @@ import (
 	"os"
 	"os/exec"
 	"path/filepath"
-	"strconv"
 	"strings"
 	"sync"
-	"sync/atomic"
 	"time"
-	"unicode"
+
 
 	"reasonix/internal/agent"
 	"reasonix/internal/billing"
@@ -36,7 +34,6 @@ import (
 	"reasonix/internal/config"
 	"reasonix/internal/diff"
 	"reasonix/internal/event"
-	"reasonix/internal/evidence"
 	"reasonix/internal/hook"
 	"reasonix/internal/i18n"
 	"reasonix/internal/jobs"
@@ -53,12 +50,6 @@ import (
 // ErrTurnRunning reports that a caller tried to start a second foreground turn
 // while one is already active in the same Controller.
 var ErrTurnRunning = errors.New("turn already running")
-
-// errNoSessionPath is returned by snapshot when a session has content to persist
-// but no resolved session path — a misconfiguration (e.g. an unresolvable data
-// dir in a bot deployment) that previously dropped conversations silently
-// (#4414). Callers log it and continue; it must never be swallowed quietly.
-var errNoSessionPath = errors.New("session has content but no session path; conversation cannot be persisted")
 
 // Controller drives one chat session. Construct with New; drive with the command
 // methods; observe through the Sink passed in Options.
@@ -445,52 +436,6 @@ func (c *Controller) recordDisplayForNewUser(startMessages int, display string) 
 
 // ckptDir derives a session's checkpoint directory from its file path
 // (…/<id>.jsonl → …/<id>.ckpt). Empty path → empty (in-memory checkpoints).
-func ckptDir(sessionPath string) string {
-	if sessionPath == "" {
-		return ""
-	}
-	return strings.TrimSuffix(sessionPath, ".jsonl") + ".ckpt"
-}
-
-// goalStatePath derives a session's persisted goal-state sidecar.
-func goalStatePath(sessionPath string) string {
-	if sessionPath == "" {
-		return ""
-	}
-	return strings.TrimSuffix(sessionPath, ".jsonl") + ".goal-state.json"
-}
-
-// rebindCheckpoints points the store at the (possibly new) session, loading any
-// checkpoints already on disk, and resets the turn boundaries. Called on
-// construction and whenever the session path changes (NewSession/Resume/SetSessionPath).
-func (c *Controller) rebindCheckpoints(sessionPath string) {
-	c.mu.Lock()
-	defer c.mu.Unlock()
-	c.goalStatePath = goalStatePath(sessionPath)
-	c.cp = checkpoint.New(ckptDir(sessionPath), c.cpRoot)
-	c.cpTurn = c.cp.NextTurn() // continue numbering past any checkpoints on disk
-	c.cpBound = c.cp.Bounds()  // rebuilt from persisted checkpoints so a resumed
-	if c.cpBound == nil {      // session can still rewind conversation / fork
-		c.cpBound = map[int]int{}
-	}
-}
-
-// beginCheckpoint opens a checkpoint for the turn about to run, recording the
-// current message count as the conversation-rewind boundary. Called at the top of
-// runTurn, before the user message is appended.
-func (c *Controller) beginCheckpoint(input string) {
-	if c.cp == nil || c.executor == nil {
-		return
-	}
-	c.mu.Lock()
-	turn := c.cpTurn
-	c.cpTurn++
-	msgIndex := len(c.executor.Session().Messages)
-	c.cpBound[turn] = msgIndex
-	c.mu.Unlock()
-	c.cp.Begin(turn, input, msgIndex)
-}
-
 // --- commands (frontend → controller) ---
 
 // runGuarded runs body on a background goroutine under a fresh cancellable
@@ -605,19 +550,6 @@ func (c *Controller) runTurnWithRaw(ctx context.Context, input, raw string) erro
 	return c.runTurnWithRawDisplay(ctx, input, raw, "")
 }
 
-func (c *Controller) runGoalLoopWithRaw(ctx context.Context, input, raw string) error {
-	return c.runGoalLoopWithRawDisplay(ctx, input, raw, "")
-}
-
-func (c *Controller) runGoalLoopWithRawDisplay(ctx context.Context, input, raw, display string) error {
-	if err := c.runTurnWithRawDisplay(ctx, input, raw, display); err != nil {
-		if ctx.Err() != nil {
-			c.stopGoal(GoalStatusStopped)
-		}
-		return err
-	}
-	return c.continueGoal(ctx)
-}
 
 func (c *Controller) runTurnWithRawDisplay(ctx context.Context, input, raw, display string) error {
 	c.maybeSessionStart(ctx)
@@ -691,296 +623,6 @@ func (c *Controller) runTurnWithRawDisplay(ctx context.Context, input, raw, disp
 	return nil
 }
 
-func (c *Controller) continueGoal(ctx context.Context) error {
-	for {
-		cont := c.advanceGoalAfterTurn()
-		if !cont {
-			return nil
-		}
-		if err := ctx.Err(); err != nil {
-			c.stopGoal(GoalStatusStopped)
-			return err
-		}
-		turn := goalContinueTurn
-		c.mu.Lock()
-		if c.goalInterceptMsg != "" {
-			turn = c.goalInterceptMsg
-			c.goalInterceptMsg = ""
-			c.mu.Unlock()
-			c.notice("goal intercept: incomplete todos remain (override with a second [goal:complete])")
-		} else {
-			c.mu.Unlock()
-		}
-		if err := c.runTurnWithRawDisplay(ctx, turn, turn, ""); err != nil {
-			if ctx.Err() != nil {
-				c.stopGoal(GoalStatusStopped)
-			}
-			return err
-		}
-	}
-}
-
-func (c *Controller) advanceGoalAfterTurn() bool {
-	reply := lastAssistantText(c.History())
-	status, reason, _ := parseGoalStatusMarker(reply)
-	var notice string
-	c.mu.Lock()
-	if strings.TrimSpace(c.goal) == "" || c.goalStatus != GoalStatusRunning {
-		c.mu.Unlock()
-		return false
-	}
-	c.goalTurns++
-	switch status {
-	case GoalStatusComplete:
-		if incomplete := c.incompleteGoalTodos(); len(incomplete) > 0 && (c.goalStrict || c.goalIntercepts == 0) {
-			// In strict mode every claim is blocked until todos are done;
-			// otherwise only the first consecutive claim is intercepted.
-			c.goalIntercepts++
-			c.goalInterceptMsg = incomplete
-			break
-		}
-		// Todos are all done — in strict mode run self-check before final
-		// completion. Non-strict mode completes immediately.
-		if c.goalStrict && !c.goalSelfCheckDone {
-			c.goalSelfCheckDone = true
-			c.goalInterceptMsg = goalSelfCheckTurn
-			break
-		}
-		// Self-check passed — complete the goal.
-		c.goalIntercepts = 0
-		c.goalSelfCheckDone = false
-		c.goalIdleTurns = 0
-		c.saveGoalState()
-		c.goal = ""
-		c.goalStatus = GoalStatusComplete
-		c.goalBlocks = 0
-		c.goalBlock = ""
-		c.goalInterceptMsg = ""
-		notice = "goal complete"
-	case GoalStatusBlocked:
-		reason = cleanGoalBlockReason(reason)
-		if reason == "" {
-			reason = "blocked"
-		}
-		if sameGoalBlock(c.goalBlock, reason) {
-			c.goalBlocks++
-		} else {
-			c.goalBlocks = 1
-			c.goalBlock = reason
-		}
-		if c.goalBlocks >= 3 {
-			c.goalStatus = GoalStatusBlocked
-			notice = "goal blocked: " + reason
-		}
-	default:
-		c.goalBlocks = 0
-		c.goalBlock = ""
-		c.goalIntercepts = 0
-		c.goalSelfCheckDone = false
-		c.goalIdleTurns = 0
-	}
-	// Idle detection: if the agent went multiple turns without any tool
-	// calls, inject a reminder to make progress (unless the goal is already
-	// completing or hitting the auto-turn limit).
-	if notice == "" && c.goalInterceptMsg == "" {
-		if c.toolWasCalledLastTurn() {
-			c.goalIdleTurns = 0
-		} else {
-			c.goalIdleTurns++
-			if c.goalIdleTurns >= maxGoalIdleTurns {
-				c.goalIdleTurns = 0
-				c.goalInterceptMsg = "No tool calls in recent turns. Either make progress with tools or signal [goal:blocked:<reason>]."
-			}
-		}
-	}
-	if notice == "" && c.goalTurns >= maxGoalAutoTurns {
-		c.goalStatus = GoalStatusBlocked
-		c.goalBlock = "goal continuation limit reached"
-		c.goalIntercepts = 0
-		c.goalSelfCheckDone = false
-		c.goalInterceptMsg = ""
-		c.goalIdleTurns = 0
-		notice = c.goalBlock
-	}
-	if notice != "" {
-		c.saveGoalState()
-	}
-	cont := notice == ""
-	c.mu.Unlock()
-	if notice != "" {
-		c.notice(notice)
-	}
-	return cont
-}
-
-// incompleteGoalTodos checks the executor's canonical todo state and evidence
-// readiness (project checks) for anything that should block [goal:complete].
-// Returns a formatted reminder string, or empty if nothing is blocking.
-func (c *Controller) incompleteGoalTodos() string {
-	if c.executor == nil {
-		return ""
-	}
-	var parts []string
-
-	// 1. Check canonical todos.
-	todos := c.executor.CanonicalTodoState()
-	if len(todos) > 0 {
-		incomplete := evidence.IncompleteTodos(todos)
-		if len(incomplete) > 0 {
-			var b strings.Builder
-			b.WriteString("the following tasks are still incomplete:")
-			for _, t := range incomplete {
-				fmt.Fprintf(&b, "\n  - %s (%s)", t.Content, t.Status)
-			}
-			parts = append(parts, b.String())
-		}
-	}
-
-	// 2. Check evidence readiness (project checks from AGENTS.md).
-	if reason := c.executor.GoalReadinessFailure(); reason != "" {
-		parts = append(parts, reason)
-	}
-
-	if len(parts) == 0 {
-		return ""
-	}
-
-	var b strings.Builder
-	b.WriteString("Goal signaled complete but issues remain:\n")
-	for _, p := range parts {
-		b.WriteString("- ")
-		b.WriteString(p)
-		b.WriteString("\n")
-	}
-	b.WriteString("Fix or use todo_write/complete_step to mark done, then [goal:complete] again.")
-	return b.String()
-}
-
-// toolWasCalledLastTurn reports whether the most recent assistant message
-// contained any tool calls, indicating the agent made observable progress.
-func (c *Controller) toolWasCalledLastTurn() bool {
-	msgs := c.History()
-	for i := len(msgs) - 1; i >= 0; i-- {
-		m := msgs[i]
-		if m.Role == provider.RoleAssistant {
-			return len(m.ToolCalls) > 0
-		}
-		if m.Role == provider.RoleUser {
-			return false
-		}
-	}
-	return false
-}
-
-func parseGoalStatusMarker(text string) (status, reason string, ok bool) {
-	lines := strings.Split(text, "\n")
-	for i := len(lines) - 1; i >= 0; i-- {
-		line := strings.TrimSpace(lines[i])
-		if line == "" {
-			continue
-		}
-		lower := strings.ToLower(line)
-		switch lower {
-		case "[goal:complete]":
-			return GoalStatusComplete, "", true
-		case "[goal:continue]":
-			return GoalStatusRunning, "", true
-		}
-		const blockedPrefix = "[goal:blocked:"
-		if strings.HasPrefix(lower, blockedPrefix) && strings.HasSuffix(line, "]") {
-			return GoalStatusBlocked, strings.TrimSpace(line[len(blockedPrefix) : len(line)-1]), true
-		}
-		return "", "", false
-	}
-	return "", "", false
-}
-
-func sameGoalBlock(a, b string) bool {
-	return normalizeGoalBlockReason(a) == normalizeGoalBlockReason(b)
-}
-
-func cleanGoalBlockReason(reason string) string {
-	return strings.Trim(strings.TrimSpace(reason), " \t\r\n:：,，.。;；!！?？-—_[]()（）")
-}
-
-func normalizeGoalBlockReason(reason string) string {
-	reason = strings.ToLower(cleanGoalBlockReason(reason))
-	var b strings.Builder
-	lastSpace := true
-	for _, r := range reason {
-		switch {
-		case unicode.IsLetter(r) || unicode.IsDigit(r):
-			b.WriteRune(r)
-			lastSpace = false
-		default:
-			if !lastSpace {
-				b.WriteByte(' ')
-				lastSpace = true
-			}
-		}
-	}
-	return strings.Join(strings.Fields(b.String()), " ")
-}
-
-func (c *Controller) stopGoal(status string) {
-	c.mu.Lock()
-	if strings.TrimSpace(c.goal) != "" && c.goalStatus == GoalStatusRunning {
-		c.goalStatus = status
-	}
-	c.goalInterceptMsg = ""
-	c.goalIntercepts = 0
-	c.goalSelfCheckDone = false
-	c.goalIdleTurns = 0
-	c.saveGoalState()
-	c.mu.Unlock()
-}
-
-// saveGoalState persists the current goal state to disk for session continuity.
-func (c *Controller) saveGoalState() {
-	if c.goalStatePath == "" || c.executor == nil {
-		return
-	}
-	todos := c.executor.CanonicalTodoState()
-	state := goalState{
-		Goal:         c.goal,
-		Status:       c.goalStatus,
-		ResearchMode: c.goalResearchMode,
-		Turns:        c.goalTurns,
-		Blocks:       c.goalBlocks,
-		Block:        c.goalBlock,
-		Strict:       c.goalStrict,
-		Todos:        todos,
-	}
-	data, err := json.Marshal(state)
-	if err != nil {
-		return
-	}
-	_ = os.MkdirAll(filepath.Dir(c.goalStatePath), 0o755)
-	_ = os.WriteFile(c.goalStatePath, data, 0o644)
-}
-
-// goalState is the serializable form of a running goal.
-type goalState struct {
-	Goal         string              `json:"goal,omitempty"`
-	Status       string              `json:"status,omitempty"`
-	ResearchMode GoalResearchMode    `json:"researchMode,omitempty"`
-	Turns        int                 `json:"turns,omitempty"`
-	Blocks       int                 `json:"blocks,omitempty"`
-	Block        string              `json:"block,omitempty"`
-	Strict       bool                `json:"strict,omitempty"`
-	Todos        []evidence.TodoItem `json:"todos,omitempty"`
-}
-
-// lastAssistantText returns the content of the most recent assistant message with
-// non-empty text — the model's final answer for the turn (its plan, in plan mode).
-func lastAssistantText(msgs []provider.Message) string {
-	for i := len(msgs) - 1; i >= 0; i-- {
-		if msgs[i].Role == provider.RoleAssistant && strings.TrimSpace(msgs[i].Content) != "" {
-			return msgs[i].Content
-		}
-	}
-	return ""
-}
 
 // Submit is the one-call entry for a simple frontend: it takes raw user input
 // and does everything — slash-command dispatch, @-reference expansion, plan-mode
@@ -1192,112 +834,6 @@ func (c *Controller) submitCommandOrTurn(trimmed, input, display string, scopedR
 	}
 }
 
-func (c *Controller) maybeAutoStartResearchGoal(input, display string) bool {
-	goal, ok := c.autoStartResearchGoalCandidate(input)
-	if !ok {
-		return false
-	}
-	if c.runner != nil {
-		displayText := display
-		if strings.TrimSpace(displayText) == "" {
-			displayText = goal
-		}
-		c.runGuarded(func(ctx context.Context) error {
-			c.SetGoalWithResearchMode(goal, GoalResearchOn)
-			c.notice(fmt.Sprintf(i18n.M.GoalSetFmt, ShortGoalForNotice(goal)))
-			block, errs := c.ResolveRefs(ctx, goal)
-			for _, e := range errs {
-				c.notice(e)
-			}
-			sent := "Start pursuing the active goal now."
-			if block != "" {
-				sent = "Referenced context:\n\n" + block + "\n\n" + sent
-			}
-			return c.runGoalLoopWithRawDisplay(ctx, sent, goal, displayText)
-		})
-	}
-	return true
-}
-
-// AutoStartResearchGoal upgrades a strong long-horizon ordinary prompt into a
-// Goal + AutoResearch run for frontends that already accepted an idle turn.
-func (c *Controller) AutoStartResearchGoal(input string) (string, bool) {
-	goal, ok := c.autoStartResearchGoalCandidate(input)
-	if !ok {
-		return "", false
-	}
-	c.SetGoalWithResearchMode(goal, GoalResearchOn)
-	c.notice(fmt.Sprintf(i18n.M.GoalSetFmt, ShortGoalForNotice(goal)))
-	return goal, true
-}
-
-func (c *Controller) autoStartResearchGoalCandidate(input string) (string, bool) {
-	goal := strings.TrimSpace(input)
-	if !shouldAutoStartResearchGoal(goal) {
-		return "", false
-	}
-	c.mu.Lock()
-	plan := c.planMode
-	running := c.running
-	activeGoal := strings.TrimSpace(c.goal) != "" && c.goalStatus == GoalStatusRunning
-	c.mu.Unlock()
-	if plan || running || activeGoal {
-		return "", false
-	}
-	return goal, true
-}
-
-func (c *Controller) rememberProjectNote(note string) {
-	if note == "" {
-		c.notice("nothing to remember")
-		return
-	}
-	if path, err := c.QuickAdd(memory.ScopeProject, note); err != nil {
-		c.notice("memory: " + err.Error())
-	} else {
-		c.notice("remembered → " + path)
-	}
-}
-
-func (c *Controller) applyGoalCommand(input, display string) bool {
-	cmd, ok := ParseGoalCommand(input)
-	if !ok {
-		return false
-	}
-	switch cmd.Action {
-	case GoalCommandSet:
-		c.SetPlanMode(false)
-		c.SetGoalWithResearchMode(cmd.Text, cmd.ResearchMode)
-		c.GoalStrict(cmd.Strict)
-		c.notice(fmt.Sprintf(i18n.M.GoalSetFmt, ShortGoalForNotice(cmd.Text)))
-		if c.runner != nil {
-			c.runGuarded(func(ctx context.Context) error {
-				return c.runGoalLoopWithRawDisplay(ctx, "Start pursuing the active goal now.", cmd.Text, display)
-			})
-		}
-	case GoalCommandClear:
-		c.ClearGoal()
-		c.notice(i18n.M.GoalCleared)
-	default:
-		goal := c.Goal()
-		if strings.TrimSpace(goal) == "" {
-			c.notice(i18n.M.GoalEmpty)
-		} else {
-			c.notice(fmt.Sprintf(i18n.M.GoalCurrentFmt, goal))
-		}
-	}
-	return true
-}
-
-func ShortGoalForNotice(goal string) string {
-	goal = strings.Join(strings.Fields(goal), " ")
-	runes := []rune(goal)
-	const max = 160
-	if len(runes) <= max {
-		return goal
-	}
-	return string(runes[:max]) + "..."
-}
 
 // applyPlanExec reads the current canonical todo list and starts a goal that
 // analyzes and dispatches independent steps concurrently via parallel_tasks.
@@ -1626,68 +1162,6 @@ func (c *Controller) Turn() int {
 // Approve answers a pending ApprovalRequest by ID: allow runs the call, session
 // also remembers a grant for the rest of the session so the same approval scope
 // is not re-prompted. Unknown/expired IDs are ignored.
-func (c *Controller) Approve(id string, allow, session, persist bool) {
-	c.mu.Lock()
-	pending := c.approvals[id]
-	delete(c.approvals, id)
-	c.mu.Unlock()
-	if pending.reply != nil {
-		pending.reply <- approvalReply{allow: allow, session: session, persist: persist} // buffered, never blocks
-	}
-}
-
-// EnableInteractiveApproval swaps the executor's gate for one that routes
-// approval decisions to the frontend via ApprovalRequest events, and wires the
-// controller in as the executor's Asker so the `ask` tool can question the user.
-// Interactive frontends (chat, desktop) call this; the headless run keeps the
-// silent gate and a nil asker from setup.
-func (c *Controller) EnableInteractiveApproval() {
-	if c.executor != nil {
-		c.executor.SetGate(c.newInteractiveGate())
-		c.executor.SetAsker(c)
-	}
-}
-
-func normalizeToolApprovalMode(mode string) string {
-	switch strings.ToLower(strings.TrimSpace(mode)) {
-	case ToolApprovalAuto, "approve", "allow":
-		return ToolApprovalAuto
-	case ToolApprovalYolo, "full", "full-access", "bypass":
-		return ToolApprovalYolo
-	default:
-		return ToolApprovalAsk
-	}
-}
-
-func (c *Controller) newInteractiveGate() *permission.Gate {
-	policy := c.policy
-	c.mu.Lock()
-	mode := normalizeToolApprovalMode(c.toolApprovalMode)
-	c.mu.Unlock()
-	switch mode {
-	case ToolApprovalAuto, ToolApprovalYolo:
-		policy.Mode = permission.Allow
-	default:
-		policy.Mode = permission.Ask
-	}
-	policy.Ask = append(policy.Ask,
-		permission.Rule{Tool: memoryRememberTool},
-		permission.Rule{Tool: memoryForgetTool},
-	)
-	gate := permission.NewGate(policy, gateApprover{c})
-	gate.OnRemember = func(rule string) {
-		if c.onRemember != nil {
-			_ = c.onRemember(rule)
-		}
-	}
-	return gate
-}
-
-func (c *Controller) refreshInteractiveGate() {
-	if c.executor != nil {
-		c.executor.SetGate(c.newInteractiveGate())
-	}
-}
 
 // Steer queues mid-turn guidance without interrupting the in-flight request.
 func (c *Controller) Steer(text string) {
@@ -1718,195 +1192,6 @@ func (c *Controller) SteerConsumed() bool {
 	return true
 }
 
-// Ask implements agent.Asker: it emits an AskRequest and blocks until
-// AnswerQuestion(ID, …) answers or ctx is cancelled. promptMu serialises it
-// against tool-approval prompts so at most one user prompt is outstanding.
-// Unlike tool-approval gates, Ask is NOT bypassed in YOLO mode — the `ask`
-// tool exists to get a genuine user decision, and YOLO only auto-approves
-// tool calls; it must not answer the user's questions for them.
-func (c *Controller) Ask(ctx context.Context, questions []event.AskQuestion) ([]event.AskAnswer, error) {
-	c.promptMu.Lock()
-	defer c.promptMu.Unlock()
-
-	c.mu.Lock()
-	c.nextID++
-	id := strconv.Itoa(c.nextID)
-	reply := make(chan []event.AskAnswer, 1)
-	c.asks[id] = pendingAsk{questions: questions, reply: reply}
-	c.mu.Unlock()
-
-	c.sink.Emit(event.Event{Kind: event.AskRequest, Ask: event.Ask{ID: id, Questions: questions}})
-
-	waitCtx, cancelWait := c.approvalWaitContext(ctx)
-	defer cancelWait()
-
-	select {
-	case ans := <-reply:
-		return ans, nil
-	case <-waitCtx.Done():
-		c.mu.Lock()
-		delete(c.asks, id)
-		c.mu.Unlock()
-		return nil, waitCtx.Err()
-	}
-}
-
-// AnswerQuestion resolves a pending AskRequest by ID with the user's selections.
-// Unknown/expired IDs are ignored.
-func (c *Controller) AnswerQuestion(id string, answers []event.AskAnswer) {
-	c.mu.Lock()
-	pending, ok := c.asks[id]
-	delete(c.asks, id)
-	c.mu.Unlock()
-	if ok {
-		pending.reply <- answers // buffered, never blocks
-	}
-}
-
-// ReplayPendingPrompts re-emits the ApprovalRequest / AskRequest event for every
-// prompt currently blocking the run loop. A frontend that reconnected or reloaded
-// after the original event has no way to rebuild its approval/ask modal otherwise,
-// so the blocked gate goroutine stays stuck forever while the session shows a
-// "waiting" status with no actionable prompt. promptMu serialises Ask and
-// requestApproval, so in practice at most one prompt is outstanding; the loops
-// stay general so a future concurrent prompt would still replay correctly.
-func (c *Controller) ReplayPendingPrompts() {
-	c.mu.Lock()
-	approvals := make([]event.Approval, 0, len(c.approvals))
-	for id, p := range c.approvals {
-		approvals = append(approvals, event.Approval{ID: id, Tool: p.tool, Subject: p.subject})
-	}
-	asks := make([]event.Ask, 0, len(c.asks))
-	for id, p := range c.asks {
-		asks = append(asks, event.Ask{ID: id, Questions: p.questions})
-	}
-	c.mu.Unlock()
-	for _, a := range approvals {
-		c.sink.Emit(event.Event{Kind: event.ApprovalRequest, Approval: a})
-	}
-	for _, a := range asks {
-		c.sink.Emit(event.Event{Kind: event.AskRequest, Ask: a})
-	}
-}
-
-// SetPlanMode flips the executor's read-only gate without touching the
-// cache-stable prompt prefix, and remembers the state so Compose can prepend the
-// plan-mode marker to outgoing turns.
-func (c *Controller) SetPlanMode(v bool) {
-	c.mu.Lock()
-	c.planMode = v
-	c.mu.Unlock()
-	if c.executor != nil {
-		c.executor.SetPlanMode(v)
-	}
-	if setter, ok := c.runner.(interface{ SetPlanMode(bool) }); ok {
-		setter.SetPlanMode(v)
-	}
-}
-
-// SetAutoPlan updates the interactive auto-plan gate for subsequent turns.
-func (c *Controller) SetAutoPlan(mode string) {
-	c.mu.Lock()
-	c.autoPlan = normalizeAutoPlan(mode)
-	c.mu.Unlock()
-}
-
-// SetReasoningLanguage updates the visible reasoning language preference for
-// subsequent turns.
-func (c *Controller) SetReasoningLanguage(lang string) {
-	mode := config.NormalizeReasoningLanguage(lang)
-	c.mu.Lock()
-	c.reasoningLanguage = mode
-	c.mu.Unlock()
-	if setter, ok := c.runner.(interface{ SetReasoningLanguage(string) }); ok {
-		setter.SetReasoningLanguage(mode)
-	} else if c.executor != nil {
-		c.executor.SetReasoningLanguage(mode)
-	}
-}
-
-// PlanMode reports whether outgoing turns currently receive the plan-mode
-// marker. Frontends use it after Compose because auto-plan may flip the mode.
-func (c *Controller) PlanMode() bool {
-	c.mu.Lock()
-	defer c.mu.Unlock()
-	return c.planMode
-}
-
-// GoalStrict enables or disables strict goal mode. In strict mode the agent
-// cannot override an incomplete-todo intercept — it must actually finish or
-// update all items before [goal:complete] is accepted.
-func (c *Controller) GoalStrict(strict bool) {
-	c.mu.Lock()
-	c.goalStrict = strict
-	c.saveGoalState()
-	c.mu.Unlock()
-}
-
-// SetGoal stores a session-scoped active goal. Compose injects it into outgoing
-// user turns, not the system prompt or tool schema, so it does not disturb the
-// cache-stable prefix.
-func (c *Controller) SetGoal(goal string) {
-	c.SetGoalWithResearchMode(goal, GoalResearchAuto)
-}
-
-func (c *Controller) SetGoalWithResearchMode(goal string, researchMode GoalResearchMode) {
-	goal = strings.TrimSpace(goal)
-	c.mu.Lock()
-	defer c.mu.Unlock()
-	if goal == "" {
-		c.goal = ""
-		c.goalStatus = GoalStatusStopped
-		c.goalResearchMode = GoalResearchAuto
-		c.goalTurns = 0
-		c.goalBlocks = 0
-		c.goalBlock = ""
-		c.goalInterceptMsg = ""
-		c.goalIntercepts = 0
-		c.goalSelfCheckDone = false
-		c.goalIdleTurns = 0
-		c.goalStrict = false
-		c.saveGoalState()
-		return
-	}
-	if c.goal == goal && c.goalStatus == GoalStatusRunning && c.goalResearchMode == researchMode {
-		return
-	}
-	c.goal = goal
-	c.goalStatus = GoalStatusRunning
-	c.goalResearchMode = researchMode
-	c.goalTurns = 0
-	c.goalBlocks = 0
-	c.goalBlock = ""
-	c.goalInterceptMsg = ""
-	c.goalIntercepts = 0
-	c.goalSelfCheckDone = false
-	c.goalIdleTurns = 0
-	c.goalStrict = false
-	c.saveGoalState()
-}
-
-func (c *Controller) ClearGoal() {
-	c.SetGoal("")
-}
-
-func (c *Controller) Goal() string {
-	c.mu.Lock()
-	defer c.mu.Unlock()
-	return c.goal
-}
-
-func (c *Controller) GoalStatus() string {
-	c.mu.Lock()
-	defer c.mu.Unlock()
-	if strings.TrimSpace(c.goal) == "" && c.goalStatus == "" {
-		return GoalStatusStopped
-	}
-	if c.goalStatus == "" {
-		return GoalStatusStopped
-	}
-	return c.goalStatus
-}
 
 // Compact runs one compaction pass on the executor's session on demand.
 // instructions is optional `/compact <focus>` guidance steering what to keep.
@@ -2017,50 +1302,6 @@ func (c *Controller) ClearSession() error {
 	return nil
 }
 
-func (c *Controller) hasUnfinishedSessionJobs(sessionPath string) bool {
-	if c.jobs == nil {
-		return false
-	}
-	return c.jobs.HasUnfinishedForSession(agent.BranchID(sessionPath))
-}
-
-func removeSessionArtifacts(path string) error {
-	if path == "" {
-		return nil
-	}
-	if err := jobs.RemoveArtifacts(path); err != nil {
-		return err
-	}
-	for _, p := range []string{path, agent.BranchMetaPath(path)} {
-		if p == "" {
-			continue
-		}
-		if err := os.Remove(p); err != nil && !os.IsNotExist(err) {
-			return err
-		}
-	}
-	if dir := ckptDir(path); dir != "" {
-		if err := os.RemoveAll(dir); err != nil && !os.IsNotExist(err) {
-			return err
-		}
-	}
-	if err := agent.DeleteSubagentsByParent(filepath.Dir(path), agent.BranchID(path)); err != nil {
-		return err
-	}
-	if err := agent.ClearCleanupPending(path); err != nil {
-		return err
-	}
-	return nil
-}
-
-// ReconcileCleanupPending retries physical cleanup for logically removed
-// sessions that were left behind by a previous process.
-func ReconcileCleanupPending(dir string) error {
-	return agent.ReconcileCleanupPending(dir, func(item agent.CleanupPendingInfo) error {
-		return removeSessionArtifacts(item.SessionPath)
-	})
-}
-
 // RewindScope selects what a Rewind restores.
 type RewindScope int
 
@@ -2071,170 +1312,6 @@ const (
 )
 
 // Checkpoints lists the session's rewind points (one per user turn), oldest first.
-func (c *Controller) Checkpoints() []checkpoint.Meta {
-	if c.cp == nil {
-		return nil
-	}
-	return c.cp.List()
-}
-
-// rewindFail emits the error as a Warn notice (so a frontend that swallows the
-// returned error — e.g. the desktop bridge's .catch — still shows the user why
-// the rewind did nothing) and returns it.
-func (c *Controller) rewindFail(err error) error {
-	c.sink.Emit(event.Event{Kind: event.Notice, Level: event.LevelWarn, Text: err.Error()})
-	return err
-}
-
-// Rewind restores the session to the start of `turn`: Code reverts every file that
-// turn (or a later one) changed to its pre-turn content; Conversation truncates the
-// message log back to that turn; Both does both. Refused while a turn is running.
-// Conversation rewind relies on the live boundary recorded at turn start, so it is
-// unavailable for turns inherited from a resumed session (code rewind still works).
-// Frontends re-render their transcript from History after the call.
-func (c *Controller) Rewind(turn int, scope RewindScope) error {
-	if c.cp == nil || c.executor == nil {
-		return c.rewindFail(fmt.Errorf("checkpoints unavailable"))
-	}
-	c.mu.Lock()
-	running := c.running
-	boundary, hasBound := c.cpBound[turn]
-	c.mu.Unlock()
-	if running {
-		return c.rewindFail(fmt.Errorf("cannot rewind while a turn is running"))
-	}
-
-	if scope == RewindCode || scope == RewindBoth {
-		written, deleted, err := c.cp.RestoreCode(turn)
-		if err != nil {
-			return c.rewindFail(fmt.Errorf("rewind code: %w", err))
-		}
-		if len(written) > 0 || len(deleted) > 0 {
-			c.sink.Emit(event.Event{Kind: event.Notice, Level: event.LevelInfo,
-				Text: fmt.Sprintf("rewound code to turn %d — %d file(s) restored, %d removed", turn, len(written), len(deleted))})
-		}
-	}
-	if scope == RewindConversation || scope == RewindBoth {
-		if !hasBound {
-			return c.rewindFail(fmt.Errorf("conversation rewind unavailable for turn %d (resumed session)", turn))
-		}
-		s := c.executor.Session()
-		// boundary is the message-log index at turn start; compaction shrinks the
-		// log without rewriting boundaries, so a stale boundary past the end means
-		// the turn was compacted away — fail loudly instead of skipping silently.
-		if boundary > len(s.Messages) {
-			return c.rewindFail(fmt.Errorf("conversation rewind unavailable for turn %d: the conversation was compacted past this point", turn))
-		}
-		s.Messages = s.Messages[:boundary]
-		c.mu.Lock()
-		c.cpTurn = turn // renumber future turns from here; later turns are gone
-		for k := range c.cpBound {
-			if k >= turn {
-				delete(c.cpBound, k)
-			}
-		}
-		c.mu.Unlock()
-		if err := c.Snapshot(); err != nil {
-			slog.Warn("controller: snapshot after rewind", "err", err)
-		}
-		c.sink.Emit(event.Event{Kind: event.Notice, Level: event.LevelInfo,
-			Text: fmt.Sprintf("rewound conversation to turn %d", turn)})
-	}
-	return nil
-}
-
-// Fork branches the conversation at the start of turn into a NEW session file,
-// preserving the current one as the branch point, and switches to the branch. Code
-// is untouched (it's a conversation operation). Like a conversation rewind it needs
-// the live boundary, so it is unavailable for resumed-session turns and refused
-// while a turn runs. Returns the new session path.
-func (c *Controller) Fork(turn int) (string, error) {
-	return c.ForkNamed(turn, "")
-}
-
-func (c *Controller) ForkNamed(turn int, name string) (string, error) {
-	return c.forkNamed(turn, name, true)
-}
-
-// ForkSession copies the conversation at the start of turn into a new session
-// file without switching this controller to it. Desktop uses this to open the
-// branch in a new tab while the source tab keeps its current transcript.
-func (c *Controller) ForkSession(turn int, name string) (string, error) {
-	return c.forkNamed(turn, name, false)
-}
-
-func (c *Controller) forkNamed(turn int, name string, switchToFork bool) (string, error) {
-	if c.executor == nil {
-		return "", c.rewindFail(fmt.Errorf("checkpoints unavailable"))
-	}
-	if c.sessionDir == "" {
-		return "", c.rewindFail(fmt.Errorf("fork needs session persistence, which is disabled"))
-	}
-	c.mu.Lock()
-	running := c.running
-	boundary, hasBound := c.cpBound[turn]
-	c.mu.Unlock()
-	if running {
-		return "", c.rewindFail(fmt.Errorf("cannot fork while a turn is running"))
-	}
-	if !hasBound {
-		return "", c.rewindFail(fmt.Errorf("fork unavailable for turn %d (resumed session)", turn))
-	}
-
-	// Persist the current conversation first so the branch point survives, then
-	// seed a fresh session with the messages up to the fork and switch to it.
-	if err := c.Snapshot(); err != nil {
-		slog.Warn("controller: pre-fork snapshot", "err", err)
-	}
-	parentPath := c.SessionPath()
-	parentID := agent.BranchID(parentPath)
-	src := c.executor.Session().Snapshot()
-	if boundary > len(src) {
-		boundary = len(src)
-	}
-	forked := append([]provider.Message(nil), src[:boundary]...)
-	sess := agent.NewSession("")
-	sess.Messages = forked
-
-	newPath := agent.NewSessionPath(c.sessionDir, c.label)
-	if err := sess.Save(newPath); err != nil {
-		return "", c.rewindFail(err)
-	}
-	if err := agent.SaveBranchMeta(newPath, agent.BranchMeta{
-		Name:             strings.TrimSpace(name),
-		ParentID:         parentID,
-		ForkTurn:         turn,
-		ForkMessageIndex: boundary,
-	}); err != nil {
-		return "", c.rewindFail(err)
-	}
-	if switchToFork {
-		c.executor.SetSession(sess)
-		c.resetPlannerSession()
-		c.mu.Lock()
-		c.sessionPath = newPath
-		c.mu.Unlock()
-		c.setActiveJobSession(newPath)
-		c.rebindCheckpoints(newPath)
-	}
-	c.sink.Emit(event.Event{Kind: event.Notice, Level: event.LevelInfo,
-		Text: fmt.Sprintf("forked conversation at turn %d into a new session", turn)})
-	return newPath, nil
-}
-
-func (c *Controller) CheckpointHasBoundary(turn int) bool {
-	c.mu.Lock()
-	defer c.mu.Unlock()
-	boundary, ok := c.cpBound[turn]
-	if !ok {
-		return false
-	}
-	// After compaction the key may still exist but the boundary value is
-	// stale (it points past the truncated message log).  Treat those
-	// turns the same as "no boundary" so the UI can disable the button.
-	return boundary <= len(c.executor.Session().Messages)
-}
-
 // Branch copies the current conversation into a child branch and switches to it.
 // Unlike Fork, it branches at the current tip and does not require a checkpoint.
 func (c *Controller) Branch(name string) (string, error) {
@@ -2372,55 +1449,6 @@ func branchDisplayName(b agent.BranchInfo) string {
 	return b.ID
 }
 
-// SummarizeFrom compresses the conversation from turn onward into one summary;
-// SummarizeUpTo compresses everything before it. Both are Claude Code's "summarize
-// from/up to here" — they restructure the message log (keeping code untouched), so
-// afterwards the per-turn boundaries no longer map and conversation rewind/fork
-// report "unavailable" until new turns rebuild them (code rewind, file-based, is
-// unaffected). Refused while a turn runs; need the live boundary.
-func (c *Controller) SummarizeFrom(ctx context.Context, turn int) error {
-	return c.summarizeAt(ctx, turn, true)
-}
-
-func (c *Controller) SummarizeUpTo(ctx context.Context, turn int) error {
-	return c.summarizeAt(ctx, turn, false)
-}
-
-func (c *Controller) summarizeAt(ctx context.Context, turn int, from bool) error {
-	if c.executor == nil {
-		return c.rewindFail(fmt.Errorf("checkpoints unavailable"))
-	}
-	c.mu.Lock()
-	running := c.running
-	boundary, hasBound := c.cpBound[turn]
-	c.mu.Unlock()
-	if running {
-		return c.rewindFail(fmt.Errorf("cannot summarize while a turn is running"))
-	}
-	if !hasBound {
-		return c.rewindFail(fmt.Errorf("summarize unavailable for turn %d (resumed session)", turn))
-	}
-	var err error
-	if from {
-		err = c.executor.SummarizeFrom(ctx, boundary)
-	} else {
-		err = c.executor.SummarizeUpTo(ctx, boundary)
-	}
-	if err != nil {
-		return c.rewindFail(err)
-	}
-	// The log was restructured; existing boundaries no longer map. Drop them (keep
-	// cpTurn monotonic so new turns don't collide with the store) — conversation
-	// rewind degrades to "unavailable" until fresh turns rebuild boundaries.
-	c.mu.Lock()
-	c.cpBound = map[int]int{}
-	c.mu.Unlock()
-	if err := c.Snapshot(); err != nil {
-		slog.Warn("controller: post-summarize snapshot", "err", err)
-	}
-	return nil
-}
-
 // Resume seeds the session from a loaded transcript and pins the active file to
 // its path so auto-save keeps appending there.
 func (c *Controller) Resume(s *agent.Session, path string) {
@@ -2478,166 +1506,6 @@ func (c *Controller) maybeColdResumePrune(path string) {
 		time.Since(last).Round(time.Minute), st.Results)})
 	if err := c.Snapshot(); err != nil {
 		slog.Warn("controller: post-prune snapshot", "err", err)
-	}
-}
-
-// Snapshot writes the executor's conversation to the active session file. No-op
-// when the executor is absent or the session has never been used (no user
-// interaction). Returns errNoSessionPath when there IS content but no resolved
-// path, so a misconfigured deployment surfaces instead of dropping data.
-// Called after every turn so a crash loses at most one in-flight prompt.
-func (c *Controller) Snapshot() error {
-	return c.snapshot(false)
-}
-
-// SnapshotActivity writes the active conversation and marks the session as
-// recently active. Use it only after a real user/model turn changes the
-// transcript; switch/close snapshots should call Snapshot so they do not reorder
-// recent-session pickers.
-func (c *Controller) SnapshotActivity() error {
-	return c.snapshot(true)
-}
-
-// midTurnSnapshotInterval is atomic (nanoseconds) so a test shrinking it
-// cannot race a previous test's still-parking autosave goroutine.
-var midTurnSnapshotInterval atomic.Int64
-
-func init() { midTurnSnapshotInterval.Store(int64(30 * time.Second)) }
-
-// autosaveWhileRunning snapshots the session periodically while a turn runs,
-// so an abrupt kill (SSH drop, force-quit) loses at most one interval of a
-// long turn instead of all of it (#3772). Session.Save copies under the lock
-// and replaces the file atomically, so racing the turn's appends is safe.
-func (c *Controller) autosaveWhileRunning(ctx context.Context) {
-	t := time.NewTicker(time.Duration(midTurnSnapshotInterval.Load()))
-	defer t.Stop()
-	for {
-		select {
-		case <-ctx.Done():
-			return
-		case <-t.C:
-			if err := c.snapshot(false); err != nil {
-				slog.Warn("controller: mid-turn snapshot", "err", err)
-			}
-		}
-	}
-}
-
-func (c *Controller) snapshot(markActivity bool) error {
-	c.mu.Lock()
-	path := c.sessionPath
-	modelRef := c.modelRef
-	c.mu.Unlock()
-	if c.executor == nil {
-		return nil
-	}
-	s := c.executor.Session()
-	if !s.HasContent() {
-		// Nothing to persist yet (e.g. a fresh session with only a system
-		// prompt) — staying quiet here is correct, not a data-loss path.
-		return nil
-	}
-	if path == "" {
-		// There IS content but nowhere to write it: this silently dropped whole
-		// bot conversations (#4414). Surface it loudly instead of returning nil
-		// so the missing session path can be diagnosed and fixed at the source.
-		slog.Warn("controller: session has content but no session path; conversation will not be persisted",
-			"label", c.Label(), "session_dir", c.SessionDir())
-		return errNoSessionPath
-	}
-	if !markActivity {
-		if _, err := agent.EnsureBranchMeta(path); err != nil {
-			return err
-		}
-	}
-	if err := s.Save(path); err != nil {
-		return err
-	}
-	if strings.TrimSpace(modelRef) != "" {
-		if err := agent.SetBranchModelPreserveUpdated(path, modelRef); err != nil {
-			return err
-		}
-	}
-	if markActivity {
-		return agent.TouchBranchMeta(path)
-	}
-	return nil
-}
-
-func (c *Controller) messageCount() int {
-	if c.executor == nil {
-		return 0
-	}
-	return len(c.executor.Session().Snapshot())
-}
-
-func (c *Controller) snapshotActivityIfChanged(startMessages int) {
-	if c.messageCount() <= startMessages {
-		return
-	}
-	if err := c.SnapshotActivity(); err != nil {
-		slog.Warn("controller: activity snapshot", "err", err)
-	}
-}
-
-// SetSessionPath pins where auto-save lands (a fresh session file minted by the
-// caller when no resume path applies).
-func (c *Controller) SetSessionPath(p string) {
-	c.mu.Lock()
-	c.sessionPath = p
-	c.mu.Unlock()
-	c.setActiveJobSession(p)
-	c.rebindCheckpoints(p)
-}
-
-// SessionDestroyHandle separates waiting for cancelled jobs from ending the
-// destroy window, so callers can move/delete persistent artifacts in between.
-type SessionDestroyHandle struct {
-	Wait    func() jobs.TeardownResult
-	WaitAll func()
-	Finish  func()
-	Async   bool
-}
-
-// BeginDestroySession marks a session as leaving active use and cancels its
-// background jobs. Call Wait before moving/deleting artifacts, then Finish after
-// persistent cleanup/move work is complete.
-func (c *Controller) BeginDestroySession(sessionPath string) SessionDestroyHandle {
-	parentSession := agent.BranchID(sessionPath)
-	if c.jobs == nil || parentSession == "" {
-		wait := func() jobs.TeardownResult { return jobs.TeardownResult{} }
-		noop := func() {}
-		return SessionDestroyHandle{Wait: wait, WaitAll: noop, Finish: noop}
-	}
-	teardown := c.jobs.BeginDestroySession(parentSession)
-	return SessionDestroyHandle{
-		Wait: func() jobs.TeardownResult {
-			return c.jobs.WaitTeardown(context.Background(), teardown, c.jobs.TeardownGrace())
-		},
-		WaitAll: func() {
-			for _, ch := range teardown.DoneChannels() {
-				<-ch
-			}
-		},
-		Finish: func() {
-			c.jobs.FinishDestroySession(parentSession)
-		},
-		Async: teardown.Async(),
-	}
-}
-
-// IsDestroyingSession reports whether sessionPath is currently in the destroy
-// window for this controller's job manager.
-func (c *Controller) IsDestroyingSession(sessionPath string) bool {
-	if c.jobs == nil {
-		return false
-	}
-	return c.jobs.IsDestroying(agent.BranchID(sessionPath))
-}
-
-func (c *Controller) setActiveJobSession(sessionPath string) {
-	if c.jobs != nil {
-		c.jobs.SetActiveSessionPath(agent.BranchID(sessionPath), sessionPath)
 	}
 }
 
@@ -2764,8 +1632,6 @@ func (c *Controller) Balance(ctx context.Context) (*billing.Balance, error) {
 
 // Host returns the running MCP host (nil when no plugins), for frontends that
 // list servers / resolve MCP prompts.
-func (c *Controller) Host() *plugin.Host { return c.host }
-
 // Commands returns the loaded custom slash commands.
 func (c *Controller) Commands() []command.Command { return c.commands }
 
@@ -2839,234 +1705,6 @@ func (c *Controller) SetSkillEnabled(name string, enabled bool) error {
 // HookRunner returns the session's hook runner (nil-safe; may hold zero hooks),
 // so a frontend can list the active hooks via `/hooks`.
 func (c *Controller) HookRunner() *hook.Runner { return c.hooks }
-
-// AddMCPServer connects an MCP server live and persists it to the config file. Its
-// tools are registered immediately and become available on the next turn (the
-// agent reads the registry per turn). The raw entry — ${VARS} intact — is what's
-// written to disk; the live connection uses the expanded form. Returns the number
-// of tools the server exposed. A save failure after a successful connect is
-// reported but non-fatal: the server still works this session.
-func (c *Controller) AddMCPServer(e config.PluginEntry) (int, error) {
-	n, err := c.connectMCPServer(e)
-	if err != nil {
-		return 0, err
-	}
-	cfg, lerr := config.Load()
-	if lerr != nil {
-		return n, fmt.Errorf("connected, but reloading config to save failed: %w", lerr)
-	}
-	if err := cfg.UpsertPlugin(e); err != nil {
-		return n, fmt.Errorf("connected, but config rejected the entry: %w", err)
-	}
-	if err := cfg.Save(); err != nil {
-		return n, fmt.Errorf("connected, but saving config failed: %w", err)
-	}
-	return n, nil
-}
-
-// ConnectMCPServer connects an MCP server entry for this session without writing
-// it to config. Desktop owns config placement so it can keep user-level settings
-// out of project reasonix.toml while preserving the CLI AddMCPServer semantics.
-func (c *Controller) ConnectMCPServer(e config.PluginEntry) (int, error) {
-	return c.connectMCPServer(e)
-}
-
-func (c *Controller) connectMCPServer(e config.PluginEntry) (int, error) {
-	exp := e.ExpandedPlugin()
-	return c.connectMCPSpec(plugin.ApplyKnownOverrides(plugin.Spec{
-		Name:    exp.Name,
-		Type:    exp.Type,
-		Command: exp.Command,
-		Args:    exp.Args,
-		Env:     exp.Env,
-		URL:     exp.URL,
-		Headers: exp.Headers,
-	}, c.WorkspaceRoot()))
-}
-
-func (c *Controller) connectMCPSpec(s plugin.Spec) (int, error) {
-	if c.host == nil {
-		c.host = plugin.NewHost()
-	}
-	tools, err := c.host.Add(c.pluginCtx, s)
-	if err != nil {
-		if !plugin.IsServerAlreadyConnected(err) {
-			return 0, err
-		}
-		toolsCtx, cancel := context.WithTimeout(c.pluginCtx, 5*time.Second)
-		defer cancel()
-		tools, err = c.host.ToolsFor(toolsCtx, s.Name)
-		if err != nil {
-			return 0, err
-		}
-	}
-	if c.reg != nil {
-		c.reg.RemovePrefix(plugin.ToolPrefix(s.Name))
-		for _, t := range tools {
-			c.reg.Add(t)
-		}
-	}
-	return len(tools), nil
-}
-
-// ImportMCPEntries persists selected MCP entries and attempts to connect them
-// live. A connection failure does not roll back the config import: the user can
-// fix local dependencies and reconnect in a later session.
-func (c *Controller) ImportMCPEntries(entries []config.PluginEntry) (total, added, updated, connected, failed, skipped int, err error) {
-	cfg, lerr := config.Load()
-	if lerr != nil {
-		return 0, 0, 0, 0, 0, 0, lerr
-	}
-	existing := make(map[string]bool, len(cfg.Plugins))
-	for _, p := range cfg.Plugins {
-		existing[p.Name] = true
-	}
-	for _, e := range entries {
-		if existing[e.Name] {
-			updated++
-		} else {
-			added++
-		}
-		if err := cfg.UpsertPlugin(e); err != nil {
-			return 0, 0, 0, 0, 0, 0, err
-		}
-		existing[e.Name] = true
-	}
-	if err := cfg.Save(); err != nil {
-		return 0, 0, 0, 0, 0, 0, err
-	}
-	for _, e := range entries {
-		if c.host != nil && containsString(c.host.ServerNames(), e.Name) {
-			skipped++
-			continue
-		}
-		if _, err := c.AddMCPServer(e); err != nil {
-			failed++
-			continue
-		}
-		connected++
-	}
-	return len(entries), added, updated, connected, failed, skipped, nil
-}
-
-func containsString(ss []string, want string) bool {
-	for _, s := range ss {
-		if s == want {
-			return true
-		}
-	}
-	return false
-}
-
-func (c *Controller) ConfiguredMCPNames() []string {
-	cfg, err := config.Load()
-	if err != nil {
-		return nil
-	}
-	names := make([]string, 0, len(cfg.Plugins))
-	for _, p := range cfg.Plugins {
-		names = append(names, p.Name)
-	}
-	return names
-}
-
-func (c *Controller) DisconnectedMCPNames() []string {
-	cfg, err := config.Load()
-	if err != nil {
-		return nil
-	}
-	connected := map[string]bool{}
-	if c.host != nil {
-		for _, name := range c.host.ServerNames() {
-			connected[name] = true
-		}
-	}
-	var names []string
-	for _, p := range cfg.Plugins {
-		if !connected[p.Name] {
-			names = append(names, p.Name)
-		}
-	}
-	return names
-}
-
-func (c *Controller) ConnectConfiguredMCPServer(name string) (int, error) {
-	cfg, err := config.Load()
-	if err != nil {
-		return 0, err
-	}
-	for _, p := range cfg.Plugins {
-		if p.Name == name {
-			return c.connectMCPServer(p)
-		}
-	}
-	return 0, fmt.Errorf("no configured MCP server named %q", name)
-}
-
-// RemoveMCPServer disconnects a live MCP server — its tools vanish from the next
-// turn — and removes it from the config file. It reports whether a live server was
-// disconnected; an error only when the name is neither connected nor in config (or
-// the config save fails). A server declared in .mcp.json disconnects for this
-// session but returns on the next start, since that file isn't ours to edit.
-func (c *Controller) RemoveMCPServer(name string) (disconnected bool, err error) {
-	if c.host != nil {
-		if prefix, ok := c.host.Remove(name); ok {
-			disconnected = true
-			if c.reg != nil {
-				c.reg.RemovePrefix(prefix)
-			}
-		}
-	}
-	cfg, lerr := config.Load()
-	if lerr != nil {
-		return disconnected, lerr
-	}
-	inConfig := cfg.RemovePlugin(name)
-	if inConfig {
-		if !disconnected && c.reg != nil {
-			c.reg.RemovePrefix(plugin.ToolPrefix(name))
-		}
-		if serr := cfg.Save(); serr != nil {
-			return disconnected, serr
-		}
-	}
-	if !disconnected && !inConfig {
-		return false, fmt.Errorf("no MCP server named %q", name)
-	}
-	return disconnected, nil
-}
-
-// DisconnectMCPServer disconnects a live server for this session without touching
-// config — the connector toggle's "off". Its tools vanish next turn; it reconnects
-// on the next session start, or now via ConnectConfiguredMCPServer (the "on").
-// Reports whether a live server was actually disconnected.
-func (c *Controller) DisconnectMCPServer(name string) bool {
-	disconnected := false
-	if c.host != nil {
-		if prefix, ok := c.host.Remove(name); ok {
-			disconnected = true
-			if c.reg != nil {
-				c.reg.RemovePrefix(prefix)
-			}
-		}
-	}
-	removedPlaceholder := 0
-	if !disconnected && c.reg != nil {
-		removedPlaceholder = c.reg.RemovePrefix(plugin.ToolPrefix(name))
-	}
-	return disconnected || removedPlaceholder > 0
-}
-
-// UnregisterMCPServerTools hides a shared MCP server from this controller only.
-// The desktop shared-host path uses this for per-tab connector toggles: the
-// shared client stays alive for sibling tabs, while this session's registry drops
-// the server's provider-visible tools before the next turn.
-func (c *Controller) UnregisterMCPServerTools(name string) bool {
-	if c.reg == nil {
-		return false
-	}
-	return c.reg.RemovePrefix(plugin.ToolPrefix(name)) > 0
-}
 
 // Label returns the human-readable model label, e.g. "deepseek-flash".
 func (c *Controller) Label() string { return c.label }
@@ -3150,767 +1788,6 @@ func (c *Controller) Jobs() []jobs.View {
 		return nil
 	}
 	return c.jobs.RunningForSession(c.parentSessionID())
-}
-
-// SetToolApprovalMode changes the runtime approval posture for permission-gated
-// tools. It does not answer business asks or plan approval.
-func (c *Controller) SetToolApprovalMode(mode string) {
-	mode = normalizeToolApprovalMode(mode)
-	var pending []chan approvalReply
-
-	c.mu.Lock()
-	c.toolApprovalMode = mode
-	c.autoApproveTools = mode == ToolApprovalYolo
-	switch mode {
-	case ToolApprovalAuto:
-		pending = c.drainApprovalsLocked(false)
-	case ToolApprovalYolo:
-		pending = c.drainApprovalsLocked(true)
-	}
-	c.mu.Unlock()
-
-	c.refreshInteractiveGate()
-	for _, reply := range pending {
-		reply <- approvalReply{allow: true}
-	}
-}
-
-func (c *Controller) ToolApprovalMode() string {
-	c.mu.Lock()
-	defer c.mu.Unlock()
-	return normalizeToolApprovalMode(c.toolApprovalMode)
-}
-
-// SetAutoApproveTools turns YOLO/full-access mode on or off for the session:
-// while on, every tool approval request is auto-allowed (writers and bash run
-// without asking). Ask requests and plan approval still reach the user. Deny
-// rules still block. Runtime-only — never written to config.
-func (c *Controller) SetAutoApproveTools(on bool) {
-	if on {
-		c.SetToolApprovalMode(ToolApprovalYolo)
-		return
-	}
-	c.SetToolApprovalMode(ToolApprovalAsk)
-}
-
-// SetBypass is the legacy name for SetAutoApproveTools. Keep it for existing
-// desktop/serve bindings and CLI code that still uses the bypass wording.
-func (c *Controller) SetBypass(on bool) {
-	c.SetAutoApproveTools(on)
-}
-
-// SetMode applies plan (read-only) and tool auto-approval together so a turn
-// submitted right after a composer mode switch can't observe a half-applied
-// gate. Turning tool auto-approval on drains any pending tool approval.
-func (c *Controller) SetMode(plan, autoApproveTools bool) {
-	c.mu.Lock()
-	c.planMode = plan
-	c.mu.Unlock()
-
-	if c.executor != nil {
-		c.executor.SetPlanMode(plan)
-	}
-	if autoApproveTools {
-		c.SetToolApprovalMode(ToolApprovalYolo)
-	} else {
-		c.SetToolApprovalMode(ToolApprovalAsk)
-	}
-}
-
-// drainApprovalsLocked removes every pending approval gate and returns their
-// reply channels; caller holds c.mu and sends {allow:true} after unlocking.
-func (c *Controller) drainApprovalsLocked(includeExplicitAsk bool) []chan approvalReply {
-	pending := make([]chan approvalReply, 0, len(c.approvals))
-	for id, approval := range c.approvals {
-		if requiresFreshApprovalTool(approval.tool) {
-			continue
-		}
-		if !includeExplicitAsk && !approval.autoDrain {
-			continue
-		}
-		delete(c.approvals, id)
-		pending = append(pending, approval.reply)
-	}
-	return pending
-}
-
-// AutoApproveTools reports whether YOLO/full-access tool auto-approval is on,
-// for status indicators and mode persistence.
-func (c *Controller) AutoApproveTools() bool {
-	return c.ToolApprovalMode() == ToolApprovalYolo
-}
-
-// Bypass is the legacy name for AutoApproveTools.
-func (c *Controller) Bypass() bool {
-	return c.AutoApproveTools()
-}
-
-// --- memory ---
-//
-// c.mem is treated as an immutable snapshot guarded by c.mu: reads take the lock
-// and return the pointer; writes mutate disk then swap in a freshly discovered
-// snapshot. A turn-tail note is queued for each write so the change applies this
-// session without disturbing the cache-stable system prefix (it folds into the
-// prefix on the next session). All of these are no-ops returning "" when memory
-// is disabled.
-
-// QuickAdd appends a one-line note to the doc-memory file for scope (project
-// REASONIX.md by default) — the write side of "#<note>". Returns the file written.
-func (c *Controller) QuickAdd(scope memory.Scope, note string) (string, error) {
-	c.mu.Lock()
-	defer c.mu.Unlock()
-	if c.mem == nil {
-		return "", nil
-	}
-	path := c.mem.DocPath(scope)
-	if path == "" {
-		return "", fmt.Errorf("no target file for memory scope %q", scope)
-	}
-	if err := memory.AppendDoc(path, note); err != nil {
-		return "", err
-	}
-	c.pendingMemory = append(c.pendingMemory, note)
-	c.refreshMemoryLocked()
-	return path, nil
-}
-
-// SaveDoc overwrites a recognized memory doc with body — the save side of the
-// desktop panel's in-place editor. Returns the file written.
-func (c *Controller) SaveDoc(path, body string) (string, error) {
-	c.mu.Lock()
-	defer c.mu.Unlock()
-	if c.mem == nil {
-		return "", nil
-	}
-	written, err := c.mem.WriteDoc(path, body)
-	if err != nil {
-		return "", err
-	}
-	// Inject the new content once on the next turn: the cached prefix still holds
-	// the pre-edit version this session, so handing the model the current text
-	// avoids a stale-guidance gap until the next session re-folds it into the
-	// prefix. Trimmed to a single tail note (drained by Compose), not per-turn.
-	c.pendingMemory = append(c.pendingMemory,
-		"Memory file "+written+" was just edited. Its current contents:\n"+strings.TrimSpace(body))
-	c.refreshMemoryLocked()
-	return written, nil
-}
-
-// SaveMemory writes an active auto-memory fact and refreshes the in-session
-// snapshot. It is the explicit user-confirmed counterpart to the model-owned
-// remember tool, used by management surfaces that preview a candidate first.
-func (c *Controller) SaveMemory(m memory.Memory) (string, error) {
-	c.mu.Lock()
-	defer c.mu.Unlock()
-	if c.mem == nil {
-		return "", nil
-	}
-	path, err := c.mem.Store.Save(m)
-	if err != nil {
-		return "", err
-	}
-	c.pendingMemory = append(c.pendingMemory,
-		"Saved memory \""+m.Name+"\": "+strings.Join(strings.Fields(m.Description), " "))
-	c.refreshMemoryLocked()
-	return path, nil
-}
-
-// ForgetMemory removes a saved auto-memory by name — the panel/TUI forget action,
-// the manual counterpart to the model's `forget` tool. It queues a turn-tail note
-// so the removal applies this session (the cached prefix still lists the fact
-// until the next session re-folds the index). The file is archived for
-// traceability by Store.Delete.
-func (c *Controller) ForgetMemory(name string) error {
-	c.mu.Lock()
-	defer c.mu.Unlock()
-	if c.mem == nil {
-		return nil
-	}
-	if err := c.mem.Store.Delete(name); err != nil {
-		return err
-	}
-	c.pendingMemory = append(c.pendingMemory,
-		"Forgot memory \""+name+"\" — disregard its line still shown in the saved-memories index until next session.")
-	c.refreshMemoryLocked()
-	return nil
-}
-
-// QueueMemory implements memory.Queue: when the model runs the remember/forget
-// tool, the tool calls this with a note that rides the next turn so the change
-// applies this session without touching the cache-stable prefix. It also
-// refreshes the snapshot a memory panel reads.
-func (c *Controller) QueueMemory(note string) {
-	c.mu.Lock()
-	defer c.mu.Unlock()
-	c.pendingMemory = append(c.pendingMemory, note)
-	c.refreshMemoryLocked()
-}
-
-// Memory returns the loaded memory snapshot (nil when memory is disabled), for
-// frontends that surface a memory panel or the /memory command. The returned
-// *Set is immutable — mutations go through QuickAdd / SaveDoc.
-func (c *Controller) Memory() *memory.Set {
-	c.mu.Lock()
-	defer c.mu.Unlock()
-	return c.mem
-}
-
-// refreshMemoryLocked re-discovers memory from disk so a later Memory() reflects
-// a just-applied write. Caller holds c.mu.
-func (c *Controller) refreshMemoryLocked() {
-	if c.mem == nil {
-		return
-	}
-	c.mem = memory.Load(memory.Options{CWD: c.mem.CWD, UserDir: c.mem.UserDir})
-}
-
-// --- approval bridge (agent gate → events) ---
-
-// gateApprover adapts the Controller to permission.Approver. It is distinct
-// from the public Approve command (different signature, different direction).
-type gateApprover struct{ c *Controller }
-
-func (g gateApprover) Approve(ctx context.Context, tool, subject string, args json.RawMessage) (bool, bool, error) {
-	subject = approvalDisplaySubject(tool, subject, args)
-	// Auto-allow without prompting while executing a just-approved plan (the plan
-	// was the approval), during an explicit continuation turn of that approved
-	// plan, or while YOLO/full-access tool auto-approval is on. Deny rules already
-	// bit before this point, so they still block.
-	g.c.mu.Lock()
-	auto := g.c.approvalBypassAllowsLocked(tool)
-	g.c.mu.Unlock()
-	if auto {
-		return true, false, nil
-	}
-	return g.c.requestApproval(ctx, tool, subject, args)
-}
-
-func approvalDisplaySubject(tool, subject string, args json.RawMessage) string {
-	switch tool {
-	case memoryRememberTool:
-		return rememberApprovalSubject(subject, args)
-	case memoryForgetTool:
-		return forgetApprovalSubject(subject, args)
-	case "move_file":
-		return moveApprovalSubject(subject, args)
-	default:
-		return subject
-	}
-}
-
-func moveApprovalSubject(fallback string, args json.RawMessage) string {
-	if len(args) == 0 {
-		return fallback
-	}
-	var in struct {
-		SourcePath      string `json:"source_path"`
-		DestinationPath string `json:"destination_path"`
-	}
-	if err := json.Unmarshal(args, &in); err != nil {
-		return fallback
-	}
-	if in.SourcePath == "" || in.DestinationPath == "" {
-		return fallback
-	}
-	return in.SourcePath + " -> " + in.DestinationPath
-}
-
-func rememberApprovalSubject(fallback string, args json.RawMessage) string {
-	if len(args) == 0 {
-		return fallback
-	}
-	var in struct {
-		Name        string `json:"name"`
-		Title       string `json:"title"`
-		Description string `json:"description"`
-		Type        string `json:"type"`
-		Body        string `json:"body"`
-	}
-	if err := json.Unmarshal(args, &in); err != nil {
-		return fallback
-	}
-	name := approvalCompactText(firstNonEmpty(in.Name, in.Title))
-	desc := approvalTruncate(approvalCompactText(in.Description), 180)
-	body := approvalTruncate(approvalCompactText(in.Body), 240)
-	typ := string(memory.NormalizeType(in.Type))
-
-	var b strings.Builder
-	b.WriteString("Save/update memory")
-	if name != "" {
-		fmt.Fprintf(&b, " %q", name)
-	}
-	if typ != "" {
-		fmt.Fprintf(&b, " [%s]", typ)
-	}
-	if desc != "" {
-		b.WriteString(": ")
-		b.WriteString(desc)
-	}
-	if body != "" {
-		if desc == "" {
-			b.WriteString(": ")
-		} else {
-			b.WriteString(" | ")
-		}
-		b.WriteString("body: ")
-		b.WriteString(body)
-	}
-	if b.Len() == len("Save/update memory") && fallback != "" {
-		return fallback
-	}
-	return b.String()
-}
-
-func forgetApprovalSubject(fallback string, args json.RawMessage) string {
-	if len(args) == 0 {
-		return fallback
-	}
-	var in struct {
-		Name string `json:"name"`
-	}
-	if err := json.Unmarshal(args, &in); err != nil {
-		return fallback
-	}
-	name := approvalCompactText(in.Name)
-	if name == "" {
-		return fallback
-	}
-	return fmt.Sprintf("Archive memory %q", name)
-}
-
-func firstNonEmpty(values ...string) string {
-	for _, value := range values {
-		if strings.TrimSpace(value) != "" {
-			return value
-		}
-	}
-	return ""
-}
-
-func approvalCompactText(s string) string {
-	return strings.Join(strings.Fields(s), " ")
-}
-
-func approvalTruncate(s string, maxRunes int) string {
-	if maxRunes <= 0 {
-		return ""
-	}
-	runes := []rune(s)
-	if len(runes) <= maxRunes {
-		return s
-	}
-	return string(runes[:maxRunes]) + "..."
-}
-
-type seedTodo struct {
-	Content string `json:"content"`
-	Status  string `json:"status"`
-	Level   int    `json:"level,omitempty"`
-}
-
-// seedPlanTodos turns an approved plan into a starter task list and emits it as a
-// synthetic todo_write event, so the live task panel populates the instant the
-// user approves — a structural guarantee, not a prompt the model might ignore.
-// The model still flips item status as it works (only it knows its own
-// progress); this just makes the list exist. No-op when the plan has no list.
-func (c *Controller) seedPlanTodos(plan string) string {
-	args := PlanTodosJSON(plan)
-	if args == "" {
-		return ""
-	}
-	t := event.Tool{ID: "plan-seed", Name: "todo_write", Args: args, ReadOnly: true}
-	c.sink.Emit(event.Event{Kind: event.ToolDispatch, Tool: t})
-	t.Output = "task list seeded from the approved plan"
-	c.sink.Emit(event.Event{Kind: event.ToolResult, Tool: t})
-	c.seedAgentTodoState(args)
-	return args
-}
-
-func (c *Controller) seedAgentTodoState(args string) {
-	if c.executor == nil {
-		return
-	}
-	todos := agentTodoStateFromArgs(args)
-	if len(todos) == 0 {
-		return
-	}
-	c.executor.SeedTodoState(todos)
-}
-
-func (c *Controller) completePlanTodos(args string) {
-	if args == "" {
-		return
-	}
-	done := completedPlanTodosJSON(args)
-	if done == "" {
-		return
-	}
-	t := event.Tool{ID: "plan-seed", Name: "todo_write", Args: done, ReadOnly: true}
-	c.sink.Emit(event.Event{Kind: event.ToolDispatch, Tool: t})
-	t.Output = "approved plan finished"
-	c.sink.Emit(event.Event{Kind: event.ToolResult, Tool: t})
-	c.replaceAgentTodoState(done)
-}
-
-func (c *Controller) replaceAgentTodoState(args string) {
-	if c.executor == nil {
-		return
-	}
-	todos := agentTodoStateFromArgs(args)
-	if len(todos) == 0 {
-		return
-	}
-	c.executor.ReplaceTodoState(todos)
-}
-
-func agentTodoStateFromArgs(args string) []evidence.TodoItem {
-	var payload struct {
-		Todos []evidence.TodoItem `json:"todos"`
-	}
-	if err := json.Unmarshal([]byte(args), &payload); err != nil {
-		return nil
-	}
-	return payload.Todos
-}
-
-// PlanTodosJSON parses an approved plan's markdown into todo_write-shaped args
-// JSON ({"todos":[...]}), or "" when the plan has no list items. The exit_plan_mode
-// path seeds via seedPlanTodos (an event); a frontend whose own approval flow
-// bypasses exit_plan_mode (the chat TUI's text-plan approval) calls this directly
-// to render the same starter checklist. Shared parsing keeps the two consistent.
-func PlanTodosJSON(plan string) string {
-	items := parsePlanTodos(plan)
-	if len(items) == 0 {
-		return ""
-	}
-	b, err := json.Marshal(map[string]any{"todos": items})
-	if err != nil {
-		return ""
-	}
-	return string(b)
-}
-
-func completedPlanTodosJSON(args string) string {
-	var p struct {
-		Todos []seedTodo `json:"todos"`
-	}
-	if err := json.Unmarshal([]byte(args), &p); err != nil || len(p.Todos) == 0 {
-		return ""
-	}
-	for i := range p.Todos {
-		p.Todos[i].Status = "completed"
-	}
-	b, err := json.Marshal(map[string]any{"todos": p.Todos})
-	if err != nil {
-		return ""
-	}
-	return string(b)
-}
-
-// parsePlanTodos extracts a starter task list from an approved plan's markdown
-// list items (bulleted or numbered): the first is in_progress, the rest pending,
-// capped so a long plan can't flood the panel. It understands ONLY markdown lists
-// — an unambiguous, standard structure — and deliberately does not guess at prose,
-// tables, or arrow sequences (those need brittle, language-specific heuristics).
-// The plan-mode marker steers the model to present its plan as a list, so this
-// catches the normal case; anything it misses is covered by the model's own
-// todo_write calls as it executes.
-func parsePlanTodos(plan string) []seedTodo {
-	var todos []seedTodo
-	for _, raw := range strings.Split(plan, "\n") {
-		item, level, ok := listItem(raw)
-		if !ok {
-			continue
-		}
-		status := "pending"
-		if len(todos) == 0 {
-			status = "in_progress"
-		}
-		todos = append(todos, seedTodo{Content: item, Status: status, Level: level})
-		if len(todos) >= 20 {
-			break
-		}
-	}
-	return todos
-}
-
-func (c *Controller) sessionMessageCount() int {
-	if c.executor == nil {
-		return 0
-	}
-	return len(c.executor.Session().Messages)
-}
-
-// hasTodoUpdateSince reports whether the model emitted its own todo_write after
-// index start, so the seeded plan todos aren't auto-completed over the model's
-// own bookkeeping.
-func (c *Controller) hasTodoUpdateSince(start int) bool {
-	if c.executor == nil {
-		return false
-	}
-	msgs := c.executor.Session().Messages
-	if start < 0 || start > len(msgs) {
-		start = len(msgs)
-	}
-	_, ok := latestTodoArgsSince(msgs, start)
-	return ok
-}
-
-func latestTodoArgsSince(msgs []provider.Message, start int) (string, bool) {
-	for i := len(msgs) - 1; i >= start; i-- {
-		for j := len(msgs[i].ToolCalls) - 1; j >= 0; j-- {
-			tc := msgs[i].ToolCalls[j]
-			if tc.Name == "todo_write" {
-				return tc.Arguments, true
-			}
-		}
-	}
-	return "", false
-}
-
-// listItem parses a markdown list line ("- x", "* x", "1. x", "2) x") into its
-// task text and a nesting level derived from leading indentation (0 for a
-// top-level item, 1 for an indented sub-step — capped at 1 since the plan is
-// two-level). ok is false when the line isn't a list item. Light inline-markdown
-// stripping keeps the checklist readable.
-func listItem(line string) (content string, level int, ok bool) {
-	trimmed := strings.TrimLeft(line, " \t")
-	if trimmed == "" {
-		return "", 0, false
-	}
-	indent := 0
-	for _, c := range line[:len(line)-len(trimmed)] {
-		if c == '\t' {
-			indent += 4
-		} else {
-			indent++
-		}
-	}
-	s := trimmed
-	// A numbered markdown heading ("### 1. Add the loader") is how models often
-	// write a phase even when asked for a list; strip the heading marker and
-	// treat it as a top-level phase. A heading without a number (a section
-	// title like "## Plan") falls through and is ignored.
-	heading := false
-	if h := strings.TrimLeft(s, "#"); h != s && strings.HasPrefix(h, " ") {
-		heading = true
-		s = strings.TrimSpace(h)
-	}
-	switch {
-	case strings.HasPrefix(s, "- "), strings.HasPrefix(s, "* "), strings.HasPrefix(s, "+ "):
-		s = s[2:]
-	default:
-		// numbered: leading digits, then "." or ")", then a space
-		i := 0
-		for i < len(s) && s[i] >= '0' && s[i] <= '9' {
-			i++
-		}
-		if i == 0 || i+1 >= len(s) || (s[i] != '.' && s[i] != ')') || s[i+1] != ' ' {
-			return "", 0, false
-		}
-		s = s[i+2:]
-	}
-	s = strings.TrimSpace(s)
-	s = strings.TrimPrefix(s, "[ ] ")
-	s = strings.TrimPrefix(s, "[x] ")
-	s = strings.ReplaceAll(s, "`", "")
-	s = strings.ReplaceAll(s, "**", "")
-	s = strings.TrimSpace(s)
-	if s == "" {
-		return "", 0, false
-	}
-	if heading {
-		return s, 0, true // a heading is always a top-level phase
-	}
-	if indent >= 2 {
-		return s, 1, true
-	}
-	return s, 0, true
-}
-
-// parseRewind parses the arguments after "/rewind". The user may provide:
-//
-//	/rewind              → latest checkpoint, both
-//	/rewind <turn>       → that turn, both
-//	/rewind <turn> <scope> → that turn, code|conversation|both
-//
-// If no turn is given, the latest checkpoint is used. If no scope is given, Both is assumed.
-func parseRewind(args string, cps []checkpoint.Meta) (int, RewindScope, error) {
-	fields := strings.Fields(args)
-	if len(fields) == 0 {
-		if len(cps) == 0 {
-			return 0, RewindBoth, fmt.Errorf("no checkpoints available")
-		}
-		return cps[len(cps)-1].Turn, RewindBoth, nil
-	}
-	turn, err := strconv.Atoi(fields[0])
-	if err != nil {
-		return 0, RewindBoth, fmt.Errorf("invalid turn: %w", err)
-	}
-	scope := RewindBoth
-	if len(fields) >= 2 {
-		switch strings.ToLower(fields[1]) {
-		case "code":
-			scope = RewindCode
-		case "conversation":
-			scope = RewindConversation
-		case "both":
-			scope = RewindBoth
-		default:
-			return 0, RewindBoth, fmt.Errorf("unknown scope %q", fields[1])
-		}
-	}
-	return turn, scope, nil
-}
-
-// approvalWaitContext returns the context the approval/ask wait blocks on. When
-// approvalTimeout is zero it just forwards ctx (interactive: wait forever). When
-// positive it layers a timeout so a headless/bot session can't hang on a prompt
-// nobody will answer (#4626, #4402); the caller treats expiry as a denial.
-func (c *Controller) approvalWaitContext(ctx context.Context) (context.Context, context.CancelFunc) {
-	if c.approvalTimeout <= 0 {
-		return ctx, func() {}
-	}
-	return context.WithTimeout(ctx, c.approvalTimeout)
-}
-
-// requestApproval emits an ApprovalRequest and blocks until Approve(ID, …)
-// answers or ctx is cancelled. A prior session grant for the same approval scope
-// short-circuits. promptMu serialises outstanding prompts.
-func (c *Controller) requestApproval(ctx context.Context, tool, subject string, args json.RawMessage) (bool, bool, error) {
-	c.mu.Lock()
-	// YOLO/full access and the just-approved-plan execution window auto-allow
-	// approval-gated tools without prompting. Plan approval is a user decision,
-	// not a tool permission, so it deliberately stays interactive.
-	if c.approvalBypassAllowsLocked(tool) || c.sessionGrantAllowsLocked(tool, subject) {
-		c.mu.Unlock()
-		return true, false, nil
-	}
-	c.mu.Unlock()
-
-	c.promptMu.Lock()
-	defer c.promptMu.Unlock()
-
-	// Re-check the grant: a session grant may have landed while we queued behind
-	// another prompt for the same subject.
-	c.mu.Lock()
-	if c.approvalBypassAllowsLocked(tool) || c.sessionGrantAllowsLocked(tool, subject) {
-		c.mu.Unlock()
-		return true, false, nil
-	}
-	c.nextID++
-	id := strconv.Itoa(c.nextID)
-	reply := make(chan approvalReply, 1)
-	c.approvals[id] = pendingApproval{tool: tool, subject: subject, autoDrain: c.autoApprovalWouldAllowLocked(tool, subject), reply: reply}
-	c.mu.Unlock()
-
-	c.sink.Emit(event.Event{Kind: event.ApprovalRequest, Approval: event.Approval{ID: id, Tool: tool, Subject: subject}})
-	if hookSubject, hookArgs, ok := permissionRequestHookPayload(tool, subject, args); ok {
-		go c.hooks.PermissionRequest(ctx, tool, hookSubject, hookArgs)
-	}
-	// The agent now needs the user's attention; a Notification hook can ping an
-	// external channel (desktop notice, phone) while the run blocks on the reply.
-	go c.hooks.Notification(ctx, approvalNotificationText(tool, subject))
-
-	waitCtx, cancelWait := c.approvalWaitContext(ctx)
-	defer cancelWait()
-
-	select {
-	case r := <-reply:
-		// Plan approvals are one-shot — never persist a session grant for them, or
-		// every future plan would auto-approve.
-		if r.allow && r.session && !requiresFreshApprovalTool(tool) {
-			rule := permission.SessionGrantRuleForScope(tool, subject)
-			c.mu.Lock()
-			c.granted[rule] = true
-			c.mu.Unlock()
-		}
-		if r.allow && r.persist && !requiresFreshApprovalTool(tool) && c.onRemember != nil {
-			c.emitRememberResult(c.onRemember(permission.RememberRuleForScope(tool, subject)))
-		}
-		return r.allow, false, nil
-	case <-waitCtx.Done():
-		c.mu.Lock()
-		delete(c.approvals, id)
-		c.mu.Unlock()
-		return false, false, waitCtx.Err()
-	}
-}
-
-func approvalNotificationText(tool, subject string) string {
-	if requiresFreshApprovalTool(tool) {
-		return "approval needed: " + tool
-	}
-	if subject == "" {
-		return "approval needed: " + tool
-	}
-	return "approval needed: " + tool + " " + subject
-}
-
-func permissionRequestHookPayload(tool, subject string, args json.RawMessage) (string, json.RawMessage, bool) {
-	switch tool {
-	case planApprovalTool:
-		return "", nil, false
-	case memoryRememberTool, memoryForgetTool:
-		return "", nil, true
-	default:
-		return subject, args, true
-	}
-}
-
-func (c *Controller) approvalBypassAllowsLocked(tool string) bool {
-	if requiresFreshApprovalTool(tool) {
-		return false
-	}
-	return c.toolApprovalMode == ToolApprovalYolo ||
-		c.approvedPlanAutoApproveTools
-}
-
-func (c *Controller) autoApprovalWouldAllowLocked(tool, subject string) bool {
-	if requiresFreshApprovalTool(tool) {
-		return false
-	}
-	policy := c.policy
-	policy.Mode = permission.Allow
-	return policy.DecideSubject(tool, false, subject) == permission.Allow
-}
-
-func (c *Controller) sessionGrantAllowsLocked(tool, subject string) bool {
-	if requiresFreshApprovalTool(tool) {
-		return false
-	}
-	for rule := range c.granted {
-		if permission.RuleMatchesString(rule, tool, subject) {
-			return true
-		}
-	}
-	return false
-}
-
-func requiresFreshApprovalTool(tool string) bool {
-	switch tool {
-	case planApprovalTool, memoryRememberTool, memoryForgetTool:
-		return true
-	default:
-		return false
-	}
-}
-
-func (c *Controller) emitRememberResult(r RememberResult) {
-	if r.Err != nil {
-		c.sink.Emit(event.Event{
-			Kind:  event.Notice,
-			Level: event.LevelWarn,
-			Text:  fmt.Sprintf(i18n.M.PermissionSaveFailedFmt, r.Rule, r.Err),
-		})
-		return
-	}
-	switch {
-	case r.Saved:
-		c.sink.Emit(event.Event{Kind: event.Notice, Level: event.LevelInfo, Text: fmt.Sprintf(i18n.M.PermissionSavedFmt, r.Path, r.Rule)})
-	case strings.TrimSpace(r.CoveredBy) != "":
-		c.sink.Emit(event.Event{Kind: event.Notice, Level: event.LevelInfo, Text: fmt.Sprintf(i18n.M.PermissionAlreadyAllowedFmt, r.Path, r.CoveredBy)})
-	}
 }
 
 // detectProjectModules scans the workspace root for top-level source directories
