+++ internal/control/mcp.go
@@ -1,0 +1,240 @@
+package control
+
+import (
+	"context"
+	"fmt"
+	"time"
+
+	"reasonix/internal/config"
+	"reasonix/internal/plugin"
+)
+
+func (c *Controller) Host() *plugin.Host { return c.host }
+
+// AddMCPServer connects an MCP server live and persists it to the config file. Its
+// tools are registered immediately and become available on the next turn (the
+// agent reads the registry per turn). The raw entry — ${VARS} intact — is what's
+// written to disk; the live connection uses the expanded form. Returns the number
+// of tools the server exposed. A save failure after a successful connect is
+// reported but non-fatal: the server still works this session.
+func (c *Controller) AddMCPServer(e config.PluginEntry) (int, error) {
+	n, err := c.connectMCPServer(e)
+	if err != nil {
+		return 0, err
+	}
+	cfg, lerr := config.Load()
+	if lerr != nil {
+		return n, fmt.Errorf("connected, but reloading config to save failed: %w", lerr)
+	}
+	if err := cfg.UpsertPlugin(e); err != nil {
+		return n, fmt.Errorf("connected, but config rejected the entry: %w", err)
+	}
+	if err := cfg.Save(); err != nil {
+		return n, fmt.Errorf("connected, but saving config failed: %w", err)
+	}
+	return n, nil
+}
+
+// ConnectMCPServer connects an MCP server entry for this session without writing
+// it to config. Desktop owns config placement so it can keep user-level settings
+// out of project reasonix.toml while preserving the CLI AddMCPServer semantics.
+func (c *Controller) ConnectMCPServer(e config.PluginEntry) (int, error) {
+	return c.connectMCPServer(e)
+}
+
+func (c *Controller) connectMCPServer(e config.PluginEntry) (int, error) {
+	exp := e.ExpandedPlugin()
+	return c.connectMCPSpec(plugin.ApplyKnownOverrides(plugin.Spec{
+		Name:    exp.Name,
+		Type:    exp.Type,
+		Command: exp.Command,
+		Args:    exp.Args,
+		Env:     exp.Env,
+		URL:     exp.URL,
+		Headers: exp.Headers,
+	}, c.WorkspaceRoot()))
+}
+
+func (c *Controller) connectMCPSpec(s plugin.Spec) (int, error) {
+	if c.host == nil {
+		c.host = plugin.NewHost()
+	}
+	tools, err := c.host.Add(c.pluginCtx, s)
+	if err != nil {
+		if !plugin.IsServerAlreadyConnected(err) {
+			return 0, err
+		}
+		toolsCtx, cancel := context.WithTimeout(c.pluginCtx, 5*time.Second)
+		defer cancel()
+		tools, err = c.host.ToolsFor(toolsCtx, s.Name)
+		if err != nil {
+			return 0, err
+		}
+	}
+	if c.reg != nil {
+		c.reg.RemovePrefix(plugin.ToolPrefix(s.Name))
+		for _, t := range tools {
+			c.reg.Add(t)
+		}
+	}
+	return len(tools), nil
+}
+
+// ImportMCPEntries persists selected MCP entries and attempts to connect them
+// live. A connection failure does not roll back the config import: the user can
+// fix local dependencies and reconnect in a later session.
+func (c *Controller) ImportMCPEntries(entries []config.PluginEntry) (total, added, updated, connected, failed, skipped int, err error) {
+	cfg, lerr := config.Load()
+	if lerr != nil {
+		return 0, 0, 0, 0, 0, 0, lerr
+	}
+	existing := make(map[string]bool, len(cfg.Plugins))
+	for _, p := range cfg.Plugins {
+		existing[p.Name] = true
+	}
+	for _, e := range entries {
+		if existing[e.Name] {
+			updated++
+		} else {
+			added++
+		}
+		if err := cfg.UpsertPlugin(e); err != nil {
+			return 0, 0, 0, 0, 0, 0, err
+		}
+		existing[e.Name] = true
+	}
+	if err := cfg.Save(); err != nil {
+		return 0, 0, 0, 0, 0, 0, err
+	}
+	for _, e := range entries {
+		if c.host != nil && containsString(c.host.ServerNames(), e.Name) {
+			skipped++
+			continue
+		}
+		if _, err := c.AddMCPServer(e); err != nil {
+			failed++
+			continue
+		}
+		connected++
+	}
+	return len(entries), added, updated, connected, failed, skipped, nil
+}
+
+func containsString(ss []string, want string) bool {
+	for _, s := range ss {
+		if s == want {
+			return true
+		}
+	}
+	return false
+}
+
+func (c *Controller) ConfiguredMCPNames() []string {
+	cfg, err := config.Load()
+	if err != nil {
+		return nil
+	}
+	names := make([]string, 0, len(cfg.Plugins))
+	for _, p := range cfg.Plugins {
+		names = append(names, p.Name)
+	}
+	return names
+}
+
+func (c *Controller) DisconnectedMCPNames() []string {
+	cfg, err := config.Load()
+	if err != nil {
+		return nil
+	}
+	connected := map[string]bool{}
+	if c.host != nil {
+		for _, name := range c.host.ServerNames() {
+			connected[name] = true
+		}
+	}
+	var names []string
+	for _, p := range cfg.Plugins {
+		if !connected[p.Name] {
+			names = append(names, p.Name)
+		}
+	}
+	return names
+}
+
+func (c *Controller) ConnectConfiguredMCPServer(name string) (int, error) {
+	cfg, err := config.Load()
+	if err != nil {
+		return 0, err
+	}
+	for _, p := range cfg.Plugins {
+		if p.Name == name {
+			return c.connectMCPServer(p)
+		}
+	}
+	return 0, fmt.Errorf("no configured MCP server named %q", name)
+}
+
+// RemoveMCPServer disconnects a live MCP server — its tools vanish from the next
+// turn — and removes it from the config file. It reports whether a live server was
+// disconnected; an error only when the name is neither connected nor in config (or
+// the config save fails). A server declared in .mcp.json disconnects for this
+// session but returns on the next start, since that file isn't ours to edit.
+func (c *Controller) RemoveMCPServer(name string) (disconnected bool, err error) {
+	if c.host != nil {
+		if prefix, ok := c.host.Remove(name); ok {
+			disconnected = true
+			if c.reg != nil {
+				c.reg.RemovePrefix(prefix)
+			}
+		}
+	}
+	cfg, lerr := config.Load()
+	if lerr != nil {
+		return disconnected, lerr
+	}
+	inConfig := cfg.RemovePlugin(name)
+	if inConfig {
+		if !disconnected && c.reg != nil {
+			c.reg.RemovePrefix(plugin.ToolPrefix(name))
+		}
+		if serr := cfg.Save(); serr != nil {
+			return disconnected, serr
+		}
+	}
+	if !disconnected && !inConfig {
+		return false, fmt.Errorf("no MCP server named %q", name)
+	}
+	return disconnected, nil
+}
+
+// DisconnectMCPServer disconnects a live server for this session without touching
+// config — the connector toggle's "off". Its tools vanish next turn; it reconnects
+// on the next session start, or now via ConnectConfiguredMCPServer (the "on").
+// Reports whether a live server was actually disconnected.
+func (c *Controller) DisconnectMCPServer(name string) bool {
+	disconnected := false
+	if c.host != nil {
+		if prefix, ok := c.host.Remove(name); ok {
+			disconnected = true
+			if c.reg != nil {
+				c.reg.RemovePrefix(prefix)
+			}
+		}
+	}
+	removedPlaceholder := 0
+	if !disconnected && c.reg != nil {
+		removedPlaceholder = c.reg.RemovePrefix(plugin.ToolPrefix(name))
+	}
+	return disconnected || removedPlaceholder > 0
+}
+
+// UnregisterMCPServerTools hides a shared MCP server from this controller only.
+// The desktop shared-host path uses this for per-tab connector toggles: the
+// shared client stays alive for sibling tabs, while this session's registry drops
+// the server's provider-visible tools before the next turn.
+func (c *Controller) UnregisterMCPServerTools(name string) bool {
+	if c.reg == nil {
+		return false
+	}
+	return c.reg.RemovePrefix(plugin.ToolPrefix(name)) > 0
+}
