+++ internal/control/memory_ops.go
@@ -1,0 +1,116 @@
+package control
+
+import (
+	"fmt"
+	"strings"
+
+	"reasonix/internal/memory"
+)
+
+func (c *Controller) QuickAdd(scope memory.Scope, note string) (string, error) {
+	c.mu.Lock()
+	defer c.mu.Unlock()
+	if c.mem == nil {
+		return "", nil
+	}
+	path := c.mem.DocPath(scope)
+	if path == "" {
+		return "", fmt.Errorf("no target file for memory scope %q", scope)
+	}
+	if err := memory.AppendDoc(path, note); err != nil {
+		return "", err
+	}
+	c.pendingMemory = append(c.pendingMemory, note)
+	c.refreshMemoryLocked()
+	return path, nil
+}
+
+// SaveDoc overwrites a recognized memory doc with body — the save side of the
+// desktop panel's in-place editor. Returns the file written.
+func (c *Controller) SaveDoc(path, body string) (string, error) {
+	c.mu.Lock()
+	defer c.mu.Unlock()
+	if c.mem == nil {
+		return "", nil
+	}
+	written, err := c.mem.WriteDoc(path, body)
+	if err != nil {
+		return "", err
+	}
+	// Inject the new content once on the next turn: the cached prefix still holds
+	// the pre-edit version this session, so handing the model the current text
+	// avoids a stale-guidance gap until the next session re-folds it into the
+	// prefix. Trimmed to a single tail note (drained by Compose), not per-turn.
+	c.pendingMemory = append(c.pendingMemory,
+		"Memory file "+written+" was just edited. Its current contents:\n"+strings.TrimSpace(body))
+	c.refreshMemoryLocked()
+	return written, nil
+}
+
+// SaveMemory writes an active auto-memory fact and refreshes the in-session
+// snapshot. It is the explicit user-confirmed counterpart to the model-owned
+// remember tool, used by management surfaces that preview a candidate first.
+func (c *Controller) SaveMemory(m memory.Memory) (string, error) {
+	c.mu.Lock()
+	defer c.mu.Unlock()
+	if c.mem == nil {
+		return "", nil
+	}
+	path, err := c.mem.Store.Save(m)
+	if err != nil {
+		return "", err
+	}
+	c.pendingMemory = append(c.pendingMemory,
+		"Saved memory \""+m.Name+"\": "+strings.Join(strings.Fields(m.Description), " "))
+	c.refreshMemoryLocked()
+	return path, nil
+}
+
+// ForgetMemory removes a saved auto-memory by name — the panel/TUI forget action,
+// the manual counterpart to the model's `forget` tool. It queues a turn-tail note
+// so the removal applies this session (the cached prefix still lists the fact
+// until the next session re-folds the index). The file is archived for
+// traceability by Store.Delete.
+func (c *Controller) ForgetMemory(name string) error {
+	c.mu.Lock()
+	defer c.mu.Unlock()
+	if c.mem == nil {
+		return nil
+	}
+	if err := c.mem.Store.Delete(name); err != nil {
+		return err
+	}
+	c.pendingMemory = append(c.pendingMemory,
+		"Forgot memory \""+name+"\" — disregard its line still shown in the saved-memories index until next session.")
+	c.refreshMemoryLocked()
+	return nil
+}
+
+// QueueMemory implements memory.Queue: when the model runs the remember/forget
+// tool, the tool calls this with a note that rides the next turn so the change
+// applies this session without touching the cache-stable prefix. It also
+// refreshes the snapshot a memory panel reads.
+func (c *Controller) QueueMemory(note string) {
+	c.mu.Lock()
+	defer c.mu.Unlock()
+	c.pendingMemory = append(c.pendingMemory, note)
+	c.refreshMemoryLocked()
+}
+
+// Memory returns the loaded memory snapshot (nil when memory is disabled), for
+// frontends that surface a memory panel or the /memory command. The returned
+// *Set is immutable — mutations go through QuickAdd / SaveDoc.
+func (c *Controller) Memory() *memory.Set {
+	c.mu.Lock()
+	defer c.mu.Unlock()
+	return c.mem
+}
+
+// refreshMemoryLocked re-discovers memory from disk so a later Memory() reflects
+// a just-applied write. Caller holds c.mu.
+func (c *Controller) refreshMemoryLocked() {
+	if c.mem == nil {
+		return
+	}
+	c.mem = memory.Load(memory.Options{CWD: c.mem.CWD, UserDir: c.mem.UserDir})
+}
